### 2022-03-22 #4263 @editorialbot
[PRE REVIEW]: DIANNA: Deep Insight And Neural Network Analysis
**Submitting author:** @elboyran (<a href="http://orcid.org/0000-0002-9834-1756">Elena Ranguelova</a>)
**Repository:** https://github.com/dianna-ai/dianna
**Branch with paper.md** (empty if default branch):
**Version:** v.0.4.0
