+++ docs/conf.py
@@ -26,7 +26,7 @@ author = u"DIANNA Team"
 # built documents.
 #
 # The short X.Y version.
-version = "0.3.0"
+version = "0.4.0"
 # The full version, including alpha/beta/rc tags.
 release = version
 
