+++ dianna/__init__.py
@@ -30,7 +30,7 @@ logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 __author__ = "DIANNA Team"
 __email__ = "dianna-ai@esciencecenter.nl"
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 
 
 def explain_image(model_or_function, input_data, method, labels=(1,), **kwargs):
