### 2022-03-24 #4263 @elboyran
@danielskatz, thank you for your message.
I would like to point out that the set of Jupyter notebooks are an integral part of the package, not only the pure Python number of lines, so they should be taken into account in the total lines of code.
