+++ .github/workflows/linting.yml
@@ -17,7 +17,7 @@ jobs:
     strategy:
       fail-fast: false
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - uses: ./.github/actions/install-python-and-package
       - name: Check style against standards using prospector
         run: prospector
