### 2022-03-29 af6ec04
simplify and streamline release GitHub workflow
Simplify: use only one command to build both the wheel and source distribution using the build module, as recommended by packaging.python.org

Streamline: only trigger the upload_test_pypi job when manually running the workflow. Before, it would also trigger on releases, but this is unnecessary. This way, the test.pypi.org channel can also be used to test an upload without having to bump the version.
