+++ .github/workflows/release.yml
@@ -7,59 +7,37 @@ on:
       - published
 
 jobs:
-  build_wheel:
-    name: Build universal wheel
+  build:
+    name: Build universal wheel and source distribution
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-
       - uses: ./.github/actions/install-python-and-package
         with:
           extras-require: publishing
-
-      - name: Build wheel
-        run: python setup.py bdist_wheel
-
+      - name: Build wheel and source distribution
+        run: python -m build
       - uses: actions/upload-artifact@v3
         with:
-          path: dist/*.whl
-
-  build_sdist:
-    name: Build source distribution
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-
-      - uses: ./.github/actions/install-python-and-package
-        with:
-          extras-require: publishing
-
-      - name: Build sdist
-        run: python setup.py sdist
-
-      - uses: actions/upload-artifact@v3
-        with:
-          path: dist/*.tar.gz
-
+          path: dist/*
 
   upload_test_pypi:
-    needs: [build_wheel, build_sdist]
+    needs: [build]
     runs-on: ubuntu-latest
+    if: github.event_name == 'workflow_dispatch'
     steps:
       - uses: actions/download-artifact@v3
         with:
           name: artifact
           path: dist
-
       - uses: pypa/gh-action-pypi-publish@v1.4.2
         with:
           user: __token__
           password: ${{ secrets.TEST_PYPI_TOKEN }}
           repository_url: https://test.pypi.org/legacy/
 
-
   upload_pypi:
-    needs: [build_wheel, build_sdist]
+    needs: [build]
     runs-on: ubuntu-latest
     if: github.event_name == 'release' && github.event.action == 'published'
     steps:
@@ -67,7 +45,6 @@ jobs:
         with:
           name: artifact
           path: dist
-
       - uses: pypa/gh-action-pypi-publish@v1.4.2
         with:
           user: __token__
