+++ dashboard/layouts.py
@@ -10,7 +10,7 @@ import base64
 import utilities
 
 #static images
-image_filename = 'logo.png' # replace with your own image
+image_filename = 'app_data/logo.png' # replace with your own image
 encoded_image = base64.b64encode(open(image_filename, 'rb').read())
 
 #rise = base64.b64encode(open('rise.png', 'rb').read())
@@ -20,10 +20,10 @@ encoded_image = base64.b64encode(open(image_filename, 'rb').read())
 colors = {
     'white': '#FFFFFF',
     'text': '#091D58',
-    'blue1' : '#063446',
+    'blue1' : '#063446', #dark blue
     'blue2' : '#0e749b',
     'blue3' : '#15b3f0',
-    'blue4' : '#d0f0fc',
+    'blue4' : '#E4F3F9', #light blue
     'yellow1' : '#f0d515'
 }
 
@@ -86,7 +86,7 @@ def get_header():
 
     return header
 
-# Nav bar
+# nav bar
 def get_navbar(p = 'images'):
 
     navbar_images = html.Div([
@@ -162,92 +162,157 @@ def get_navbar(p = 'images'):
     else:
         return navbar_text
 
+# uploads bar        
 def get_uploads():
 
+    # uploads row
     uploads = html.Div([
 
-        html.Div(['b'],
-            className = 'three columns',
-            style = {'color' : colors['blue4']}),
-
+        # uploads first col (3-col)
         html.Div([
-            dcc.Upload(
-                id='upload-image',
-                children=html.Div([
-                    'Drag and Drop or ',
-                    html.A('Select Image')
-                ]),
-                style={
-                    'width': '80%',
-                    'height': '40px',
-                    'lineHeight': '40px',
-                    'borderWidth': '1px',
-                    'borderStyle': 'dashed',
-                    'borderRadius': '3px',
-                    'textAlign': 'center',
-                    'align-items': 'center',
-                    'margin': '10px',
-                    'color' : colors['blue1']
-                },
-                # Allow multiple files to be uploaded
-                multiple=True
+
+            # select image row
+            html.Div([
+                dcc.Upload(
+                    id='upload-image',
+                    children=html.Div([
+                        'Drag and Drop or ',
+                        html.A('Select Image')
+                    ]),
+                    style={
+                        'width': '80%',
+                        'height': '40px',
+                        'lineHeight': '40px',
+                        'borderWidth': '1px',
+                        'borderStyle': 'dashed',
+                        'borderRadius': '3px',
+                        'margin-left': '30px',
+                        'margin-top': '20px',
+                        'color' : colors['blue1']
+                    },
+                    # Allow multiple files to be uploaded
+                    multiple=True
+                )
+            ],
+            className = 'row',
             ),
-            html.Div(id='output-image-upload',
+
+            # plot selected image row
+            html.Div(dcc.Graph(
+                    id='graph_test',
+                    figure = utilities.blank_fig()
+                    ),
+            className = 'row',
             style = {
-                'background-color' : colors['blue4'],
-                'textAlign': 'center',
-                'height': '230px'}),
-            ], className = 'three columns', style = {'align-items': 'center'}),
-
-        html.Div([
-            dcc.Upload(
-                id='upload-model',
-                children=html.Div([
-                    'Drag and Drop or ',
-                    html.A('Select Model')
-                ]),
-                style={
-                    'width': '80%',
-                    'height': '40px',
-                    'lineHeight': '40px',
-                    'borderWidth': '1px',
-                    'borderStyle': 'dashed',
-                    'borderRadius': '3px',
-                    'textAlign': 'center',
-                    'align-items': 'center',
-                    'margin': '10px',
-                    'color' : colors['blue1']
-                },
[…]
