+++ dashboard/layouts.py
@@ -1,20 +1,10 @@
-import plotly.express as px
-import plotly.graph_objects as go
-from plotly.subplots import make_subplots
-from jupyter_dash import JupyterDash
 from dash import html, dcc
-from dash.dependencies import Input, Output
-import dash_bootstrap_components as dbc
-import pandas as pd
 import base64
 import utilities
 
 #static images
 image_filename = 'app_data/logo.png' # replace with your own image
 encoded_image = base64.b64encode(open(image_filename, 'rb').read())
-
-#rise = base64.b64encode(open('rise.png', 'rb').read())
-#kernels = base64.b64encode(open('kernels.png', 'rb').read())
 
 # colors
 colors = {
@@ -204,7 +194,7 @@ def get_uploads():
                     ),
             className = 'row',
             style = {
-                #'height': '230px'
+                'height': '230px'
                 }
             ), 
         
@@ -238,7 +228,11 @@ def get_uploads():
             html.Div(id='output-model-upload',
             className = 'row',
             style = {
-                'height': '230px'}
+                'height': '230px',
+                'margin-top': '20px',
+                #'margin-bottom': '50px',
+                #'fontSize': 8,
+                'color' : colors['blue1']}
             )
 
             ],
@@ -296,75 +290,3 @@ def get_uploads():
         })
 
     return uploads
-
-# def get_method_sel():
-
-#     method_sel = html.Div([
-
-#         html.Div(['b'],
-#             className = 'five columns',
-#             style = {'color' : colors['blue4']}),
-#         html.Div([
-#             dcc.Dropdown(id = 'method_sel2',
-#                 options = [{'label': 'RISE', 'value': 'RISE'},
-#                            {'label': 'KernelSHAP', 'value': 'KernelSHAP'},
-#                            {'label': 'LIME', 'value': 'LIME'}],
-#                 placeholder = "Select one/more methods",
-#                 value=[],
-#                 multi = True
-#             ),
-            #html.Button(
-            #    id='submit-val',
-            #    children = html.Div(['Get explanation']), n_clicks=0,
-            #    style={
-            #        'width': '100%',
-            #        #'height': '40px',
-            #        #'lineHeight': '40px',
-            #        'borderWidth': '1px',
-            #        #'borderStyle': 'dashed',
-            #        'borderRadius': '3px',
-            #        'textAlign': 'center',
-            #        'align-items': 'center',
-            #        'margin': '10px',
-            #        'color' : colors['white'],
-            #        'background-color' : colors['blue2']}
-            #    )
-            # ], 
-    #         className = 'two columns')
-    # ], className = 'row', style = {
-    #     'textAlign': 'center',
-    #     'background-color' : colors['blue4'],
-    #     'align-items': 'center'})
-
-    # return method_sel
-
-# def get_pred():
-
-#     pred = html.Div([
-#         html.Div(['b'],
-#             className = 'five columns',
-#             style = {'color' : colors['blue4']}),
-#         html.Div(id='output-state', className = 'two columns')
-#         ], className = 'row', style = {
-#         'textAlign': 'center',
-#         'background-color' : colors['blue4'],
-#         'align-items': 'center'})
-
-#     return pred
-
-# def get_method_plot():
-
-#     method_plot = html.Div([
-#         html.Div(['b'],
-#             className = 'five columns',
-#             style = {'color' : colors['blue4']}),
-#         html.Div([dcc.Graph(id='rise', figure = utilities.blank_fig())], className = 'two columns'),
-#         ],
-#         className="row",
-#         style={
-#             'background-color' : colors['blue4'],
-#             'textAlign': 'center',
-#             'align-items': 'center',
-#             'verical-align': 'center'})
-
-#     return method_plot
