### 2022-04-05 5feaf4f
add text page to dashboard
