+++ dashboard/callbacks.py
@@ -1,0 +1,407 @@
+# Plotly
+import plotly.graph_objects as go
+from plotly.subplots import make_subplots
+# Dash&Flask
+from jupyter_dash import JupyterDash
+import dash
+from dash import html
+from dash.exceptions import PreventUpdate
+from flask_caching import Cache
+# Onnx
+import onnx
+from onnx_tf.backend import prepare
+# Others
+import dianna
+from dianna import visualization
+from dianna import utils
+from scipy.special import expit as sigmoid
+import spacy
+from torchtext.data import get_tokenizer
+from torchtext.vocab import Vectors
+import os
+import base64
+import layouts
+import utilities
+import numpy as np
+import dianna
+import warnings
+warnings.filterwarnings('ignore') # disable warnings relateds to versions of tf
+
+folder_on_server = "app_data"
+os.makedirs(folder_on_server, exist_ok=True)
+
+# Build App
+external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
+app = JupyterDash(
+    __name__,
+    external_stylesheets=external_stylesheets,
+    prevent_initial_callbacks=True,
+    suppress_callback_exceptions=True)
+
+# Caching
+cache = Cache(app.server, config={
+    'CACHE_TYPE': 'filesystem',
+    'CACHE_DIR': 'cache'
+})
+cache.clear()
+
+# global variables
+class_name_mnist = ['digit 0', 'digit 1']
+class_name_text = ("negative", "positive")
+
+try:
+    spacy.load("en_core_web_sm")
+except: # If not present, we download
+    spacy.cli.download("en_core_web_sm")
+    spacy.load("en_core_web_sm")
+
+@app.callback(dash.dependencies.Output('output-model-img-upload', 'children'),
+              dash.dependencies.Input('upload-model-img', 'contents'),
+              dash.dependencies.State('upload-model-img', 'filename'))
+def upload_model(contents, filename):
+    if contents is not None:
+        try:
+            if 'onnx' in filename[0]:
+
+                content_type, content_string = contents[0].split(',')
+
+                with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
+                    f.write(base64.b64decode(content_string))
+
+                return html.Div([f'{filename[0]} uploaded'])
+            else:
+                return html.Div([
+                    html.P('File format error!'),
+                    html.Br(),
+                    html.P('Please upload only models in .onnx format.')
+                    ])
+        except Exception as e:
+            print(e)
+            return html.Div(['There was an error processing this file.'])
+    else:
+        raise PreventUpdate
+
+@app.callback(dash.dependencies.Output('output-model-text-upload', 'children'),
+              dash.dependencies.Input('upload-model-text', 'contents'),
+              dash.dependencies.State('upload-model-text', 'filename'))
+def upload_model(contents, filename):
+    if contents is not None:
+        try:
+            if 'onnx' in filename[0]:
+
+                content_type, content_string = contents[0].split(',')
+
+                with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
+                    f.write(base64.b64decode(content_string))
+
+                return html.Div([f'{filename[0]} uploaded'])
+            else:
+                return html.Div([
+                    html.P('File format error!'),
+                    html.Br(),
+                    html.P('Please upload only models in .onnx format.')
+                    ])
+        except Exception as e:
+            print(e)
+            return html.Div(['There was an error processing this file.'])
+    else:
+        raise PreventUpdate
+
+@app.callback(dash.dependencies.Output('graph_test', 'figure'),
+              dash.dependencies.Input('upload-image', 'contents'),
[…]
