### 2022-04-06 1f19183
add functionalities to text page
