+++ dashboard/callbacks.py
@@ -12,16 +12,12 @@ import onnx
 from onnx_tf.backend import prepare
 # Others
 import dianna
-from dianna import visualization
-from dianna import utils
-from scipy.special import expit as sigmoid
 import spacy
-from torchtext.data import get_tokenizer
-from torchtext.vocab import Vectors
 import os
 import base64
 import layouts
 import utilities
+from utilities import MovieReviewsModelRunner, _create_html, _highlight_word
 import numpy as np
 import dianna
 import warnings
@@ -55,58 +51,9 @@ except: # If not present, we download
     spacy.cli.download("en_core_web_sm")
     spacy.load("en_core_web_sm")
 
-@app.callback(dash.dependencies.Output('output-model-img-upload', 'children'),
-              dash.dependencies.Input('upload-model-img', 'contents'),
-              dash.dependencies.State('upload-model-img', 'filename'))
-def upload_model(contents, filename):
-    if contents is not None:
-        try:
-            if 'onnx' in filename[0]:
+########################### Images page ###########################
 
-                content_type, content_string = contents[0].split(',')
-
-                with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
-                    f.write(base64.b64decode(content_string))
-
-                return html.Div([f'{filename[0]} uploaded'])
-            else:
-                return html.Div([
-                    html.P('File format error!'),
-                    html.Br(),
-                    html.P('Please upload only models in .onnx format.')
-                    ])
-        except Exception as e:
-            print(e)
-            return html.Div(['There was an error processing this file.'])
-    else:
-        raise PreventUpdate
-
-@app.callback(dash.dependencies.Output('output-model-text-upload', 'children'),
-              dash.dependencies.Input('upload-model-text', 'contents'),
-              dash.dependencies.State('upload-model-text', 'filename'))
-def upload_model(contents, filename):
-    if contents is not None:
-        try:
-            if 'onnx' in filename[0]:
-
-                content_type, content_string = contents[0].split(',')
-
-                with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
-                    f.write(base64.b64decode(content_string))
-
-                return html.Div([f'{filename[0]} uploaded'])
-            else:
-                return html.Div([
-                    html.P('File format error!'),
-                    html.Br(),
-                    html.P('Please upload only models in .onnx format.')
-                    ])
-        except Exception as e:
-            print(e)
-            return html.Div(['There was an error processing this file.'])
-    else:
-        raise PreventUpdate
-
+# uploading test image
 @app.callback(dash.dependencies.Output('graph_test', 'figure'),
               dash.dependencies.Input('upload-image', 'contents'),
               dash.dependencies.State('upload-image', 'filename'))
@@ -154,20 +101,32 @@ def upload_image(contents, filename):
             return utilities.blank_fig(
                     text='There was an error processing this file.')
 
+# uploading model for image
+@app.callback(dash.dependencies.Output('output-model-img-upload', 'children'),
+              dash.dependencies.Input('upload-model-img', 'contents'),
+              dash.dependencies.State('upload-model-img', 'filename'))
+def upload_model(contents, filename):
+    if contents is not None:
+        try:
+            if 'onnx' in filename[0]:
 
-@app.callback(dash.dependencies.Output('text_test', 'children'),
-              dash.dependencies.Input('submit-text', 'n_clicks'),
-              dash.dependencies.State('upload-text', 'value'))
-def upload_text(clicks, input_value):
-    if clicks is not None:
-        return html.Div([
-                    html.P('Input string for the model is:'),
+                content_type, content_string = contents[0].split(',')
+
[…]
