### 2022-04-06 #4263 @Kevin-Mattheus-Moerman
@elboyran the board has reviewed this work and have determined it may be in scope for JOSS. I have therefore removed the `query-scope` label and we will now proceed to find a handling editor.
