### 2022-04-07 36f4711
add functionalities to text page
