+++ dashboard/callbacks.py
@@ -1,6 +1,7 @@
 # Plotly
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
+import plotly.express as px
 # Dash&Flask
 from jupyter_dash import JupyterDash
 import dash
@@ -11,13 +12,15 @@ from flask_caching import Cache
 import onnx
 from onnx_tf.backend import prepare
 # Others
+from PIL import Image
+from html2image import Html2Image
 import dianna
 import spacy
 import os
 import base64
 import layouts
 import utilities
-from utilities import MovieReviewsModelRunner, _create_html, _highlight_word
+from utilities import MovieReviewsModelRunner, _create_html
 import numpy as np
 import dianna
 import warnings
@@ -43,7 +46,7 @@ cache.clear()
 
 # global variables
 class_name_mnist = ['digit 0', 'digit 1']
-class_name_text = ("negative", "positive")
+class_name_text = ["negative", "positive"]
 
 try:
     spacy.load("en_core_web_sm")
@@ -188,7 +191,7 @@ def compute_value_i(method_sel, fn_m, fn_i):
 # update image explainations
 @app.callback(
     dash.dependencies.Output('output-state-img', 'children'),
-    dash.dependencies.Output('graph', 'figure'),
+    dash.dependencies.Output('graph_img', 'figure'),
     dash.dependencies.State("upload-model-img", "filename"),
     dash.dependencies.State("upload-image", "filename"),
     dash.dependencies.Input("signal_image", "data"),
@@ -347,13 +350,29 @@ def upload_model(contents, filename):
 # and for all time.
 @cache.memoize()
 def global_store_t(method_sel, model_runner, input_text):
+
+    predictions = model_runner(input_text)
+    class_name = [c for c in class_name_text]
+    pred_class = class_name[np.argmax(predictions)]
+    labels = tuple(class_name_text)
+    pred_idx = labels.index(pred_class)
+
     # expensive query
     if method_sel == "LIME":
+
         relevances = dianna.explain_text(
             model_runner,
             input_text,
             method_sel,
-            label=tuple(class_name_text)
+            label=pred_idx
+            )
+
+    elif method_sel == "RISE":
+        relevances = dianna.explain_text(
+            model_runner,
+            input_text,
+            method_sel, # RISE
+            labels=[pred_idx]
             )
 
     return relevances
@@ -376,18 +395,17 @@ def compute_value_t(method_sel, fn_m, input_text):
 
         for m in method_sel:
             # compute value and send a signal when done
-
             try:
                 global_store_t(m, model_runner, input_text)
             except Exception:
-                return method_sel
-                
+                return method_sel     
         return method_sel
 
 # update text explainations
 @app.callback(
     dash.dependencies.Output("output-state-text", "children"),
-    dash.dependencies.Output("text_expl", "children"),
+    dash.dependencies.Output("graph_text_lime", "figure"),
+    dash.dependencies.Output("graph_text_rise", "figure"),
     dash.dependencies.State("upload-model-text", "filename"),
     dash.dependencies.State("upload-text", "value"),
     dash.dependencies.Input("signal_text", "data"),
@@ -400,9 +418,9 @@ def update_multi_options_t(fn_m, input_text, sel_methods, new_model, new_text):
 
     if (ctx.triggered[0]["prop_id"] == "upload-model-text.filename") or (ctx.triggered[0]["prop_id"] == "upload-text.value") or (not ctx.triggered):
         cache.clear()
-        return html.Div(['']), html.Div([''])
+        return html.Div(['']), utilities.blank_fig(), utilities.blank_fig()
     elif (not sel_methods):
-        return html.Div(['']), html.Div([''])
[…]
