### 2022-04-12 6ebbeec
fix prospector checks
