+++ dashboard/callbacks.py
@@ -67,7 +67,7 @@ def upload_image(contents, filename):
 
             if 'jpg' in filename[0]:
 
-                content_type, content_string = contents[0].split(',')
+                _, content_string = contents[0].split(',')
 
                 with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
                     f.write(base64.b64decode(content_string))
@@ -95,9 +95,9 @@ def upload_image(contents, filename):
                 fig.layout.paper_bgcolor = layouts.colors['blue4']
 
                 return fig
-            else:
-                return utilities.blank_fig(
-                    text='File format error! <br><br>Please upload only images in .jpg format.')
+
+            return utilities.blank_fig(
+                text='File format error! <br><br>Please upload only images in .jpg format.')
         
         except Exception as e:
             print(e)
@@ -110,24 +110,24 @@ def upload_image(contents, filename):
 @app.callback(dash.dependencies.Output('output-model-img-upload', 'children'),
               dash.dependencies.Input('upload-model-img', 'contents'),
               dash.dependencies.State('upload-model-img', 'filename'))
-def upload_model(contents, filename):
+def upload_model_img(contents, filename):
     '''Takes in the model file, returns a print statement about its uploading state'''
     if contents is not None:
         try:
             if 'onnx' in filename[0]:
 
-                content_type, content_string = contents[0].split(',')
+                _, content_string = contents[0].split(',')
 
                 with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
                     f.write(base64.b64decode(content_string))
 
                 return html.Div([f'{filename[0]} uploaded'])
-            else:
-                return html.Div([
-                    html.P('File format error!'),
-                    html.Br(),
-                    html.P('Please upload only models in .onnx format.')
-                    ])
+
+            return html.Div([
+                html.P('File format error!'),
+                html.Br(),
+                html.P('Please upload only models in .onnx format.')
+                ])
         except Exception as e:
             print(e)
             return html.Div(['There was an error processing this file.'])
@@ -146,7 +146,7 @@ def global_store_i(method_sel, model_path, image_test):
     if method_sel == "RISE":
         relevances = dianna.explain_image(
             model_path, image_test, method=method_sel,
-            labels=[i for i in range(2)],
+            labels=list(range(2)),
             n_masks=5000, feature_res=8, p_keep=.1,
             axis_labels=('height','width','channels'))
 
@@ -162,7 +162,7 @@ def global_store_i(method_sel, model_path, image_test):
             model_path, image_test * 256, 'LIME',
             axis_labels=('height','width','channels'),
             random_state=2,
-            labels=[i for i in range(2)],
+            labels=list(range(2)),
             preprocess_function=utilities.preprocess_function)
 
     return relevances
@@ -180,20 +180,20 @@ def compute_value_i(method_sel, fn_m, fn_i):
 
     if (method_sel is None) or (fn_m is None) or (fn_i is None):
         raise PreventUpdate
-    else:
-        for m in method_sel:
-            # compute value and send a signal when done
-            data_path = os.path.join(folder_on_server, fn_i[0])
-            image_test = utilities.open_image(data_path)
 
-            model_path = os.path.join(folder_on_server, fn_m[0])
+    for m in method_sel:
+        # compute value and send a signal when done
+        data_path = os.path.join(folder_on_server, fn_i[0])
+        image_test = utilities.open_image(data_path)
 
-            try:
-                global_store_i(m, model_path, image_test)
-            except Exception:
-                return method_sel
-                
-        return method_sel
+        model_path = os.path.join(folder_on_server, fn_m[0])
+
+        try:
[…]
