### 2022-04-12 3dedf03
Update standup-checklist.md
