### 2022-04-13 3956fec
Add Tokenizer class, with default word based one, plus notebook with some tests.
