### 2022-04-14 ebdd952
Change text_length in rise to more accurate num_tokens
