### 2022-04-14 5431225
Add _convert_tokens_to_string method to RISE.
