### 2022-04-14 603e6d5
Move mask_string from RISE __init__ to explain_text method
