+++ dianna/methods/rise.py
@@ -1,3 +1,4 @@
+from tkinter import W
 import numpy as np
 from skimage.transform import resize
 from tqdm import tqdm
@@ -19,7 +20,7 @@ class RISE:
     required_labels = ('channels', )
 
     def __init__(self, n_masks=1000, feature_res=8, p_keep=None,  # pylint: disable=too-many-arguments
-                 axis_labels=None, preprocess_function=None, mask_string="UNKWORDZ"):
+                 axis_labels=None, preprocess_function=None):
         """RISE initializer.
 
         Args:
@@ -30,7 +31,6 @@ class RISE:
                                                If a list, the name of each axis where the index
                                                in the list is the axis index
             preprocess_function (callable, optional): Function to preprocess input data with
-            mask_string (str, optional): String to replace masked tokens with (text only)
         """
         self.n_masks = n_masks
         self.feature_res = feature_res
@@ -39,9 +39,8 @@ class RISE:
         self.masks = None
         self.predictions = None
         self.axis_labels = axis_labels if axis_labels is not None else []
-        self.mask_string = mask_string
 
-    def explain_text(self, model_or_function, input_text, labels=(0,), batch_size=100):
+    def explain_text(self, model_or_function, input_text, labels=(0,), batch_size=100, mask_string="UNKWORDZ"):
         """Runs the RISE explainer on text.
 
            The model will be called with masked versions of the input text.
@@ -52,6 +51,7 @@ class RISE:
             input_text (np.ndarray): Text to be explained
             labels (list(int)): Labels to be explained
             batch_size (int): Batch size to use for running the model.
+            mask_string (str, optional): String to replace masked tokens with (defaults to "UNKWORDZ")
 
         Returns:
             Explanation heatmap for each class (np.ndarray).
@@ -59,30 +59,30 @@ class RISE:
         runner = utils.get_function(model_or_function, preprocess_function=self.preprocess_function)
         input_tokens = np.asarray(model_or_function.tokenizer(input_text))
         num_tokens = len(input_tokens)
-        active_p_keep = self._determine_p_keep_for_text(input_tokens, runner) if self.p_keep is None else self.p_keep
+        active_p_keep = self._determine_p_keep_for_text(input_tokens, runner, mask_string) if self.p_keep is None else self.p_keep
         input_shape = (num_tokens,)
         self.masks = self._generate_masks_for_text(input_shape, active_p_keep,
                                                    self.n_masks)  # Expose masks for to make user inspection possible
-        masked_sentences = self._create_masked_sentences(input_tokens, self.masks)
+        masked_sentences = self._create_masked_sentences(input_tokens, self.masks, mask_string)
         saliencies = self._get_saliencies(runner, masked_sentences, num_tokens, batch_size, active_p_keep)
         return self._reshape_result(input_tokens, labels, saliencies)
 
-    def _determine_p_keep_for_text(self, input_data, runner, n_masks=100):
+    def _determine_p_keep_for_text(self, input_data, runner, mask_string, n_masks=100):
         """See n_mask default value https://github.com/dianna-ai/dianna/issues/24#issuecomment-1000152233."""
         p_keeps = np.arange(0.1, 1.0, 0.1)
         stds = []
         for p_keep in p_keeps:
-            std = self._calculate_mean_class_std_for_text(p_keep, runner, input_data, n_masks=n_masks)
+            std = self._calculate_mean_class_std_for_text(p_keep, runner, input_data, mask_string, n_masks=n_masks)
             stds += [std]
         best_i = np.argmax(stds)
         best_p_keep = p_keeps[best_i]
         print(f'Rise parameter p_keep was automatically determined at {best_p_keep}')
         return best_p_keep
 
[…]
