### 2022-04-14 e014749
Replace WordTokenizer with SpacyTokenizer
