+++ dianna/utils/tokenizers.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 from typing import List
-import string
+from torchtext.data import get_tokenizer
 
 
 class Tokenizer(ABC):
@@ -26,14 +26,13 @@ class Tokenizer(ABC):
         pass
 
 
-class WordBasedTokenizer(Tokenizer):
-    def __init__(self, mask_token: str = "UNKWORDZ"):
+class SpacyTokenizer(Tokenizer):
+    def __init__(self, name: str = 'en_core_web_sm', mask_token: str = "UNKWORDZ"):
         super().__init__(mask_token)
+        self.spacy_tokenizer = get_tokenizer('spacy', name)
     
     def tokenize(self, sentence: str) -> List[str]:
-        """Strip punctuation and split at spaces."""
-        sentence_nopunc = sentence.translate(str.maketrans('', '', string.punctuation))
-        tokens = sentence_nopunc.split(' ')
+        tokens = self.spacy_tokenizer(sentence)
         return tokens
          
     def convert_tokens_to_string(self, tokens: List[str]) -> str:
