### 2022-04-14 291912b
Remove mask_string, using instead model_or_function's tokenizer
