### 2022-04-14 31d6fc5
Remove RISE's _convert_tokens_to_strings, using instead the tokenizer's method.
