### 2022-04-20 dad11db
add new WIP notebook for RISE-style embedding explainer
