### 2022-04-21 9c2e921
add set_all_the_seeds function
