### 2022-04-22 9d21408
Fix prospector errors regarding docs
