+++ dashboard/callbacks.py
@@ -60,7 +60,7 @@ except Exception: # If not present, we download
               dash.dependencies.Input('upload-image', 'contents'),
               dash.dependencies.State('upload-image', 'filename'))
 def upload_image(contents, filename):
-    '''Takes in test image file, returns it as a Plotly figure'''
+    """Takes in test image file, returns it as a Plotly figure."""
     if contents is not None:
 
         try:
@@ -111,7 +111,7 @@ def upload_image(contents, filename):
               dash.dependencies.Input('upload-model-img', 'contents'),
               dash.dependencies.State('upload-model-img', 'filename'))
 def upload_model_img(contents, filename):
-    '''Takes in the model file, returns a print statement about its uploading state'''
+    """Takes in the model file, returns a print statement about its uploading state."""
     if contents is not None:
         try:
             if 'onnx' in filename[0]:
@@ -140,8 +140,7 @@ def upload_model_img(contents, filename):
 # and for all time.
 @cache.memoize()
 def global_store_i(method_sel, model_path, image_test):
-    '''Takes in the selected XAI method, the model path and the image to test,
-    returns the explainations array'''
+    """Takes in the selected XAI method, the model path and the image to test, returns the explainations array."""
     # expensive query
     if method_sel == "RISE":
         relevances = dianna.explain_image(
@@ -175,9 +174,7 @@ def global_store_i(method_sel, model_path, image_test):
     dash.dependencies.State("upload-image", "filename"),
     ])
 def compute_value_i(method_sel, fn_m, fn_i):
-    '''Takes in the selected XAI method, the model and the image filenames,
-    returns the selected XAI method'''
-
+    """Takes in the selected XAI method, the model and the image filenames, returns the selected XAI method."""
     if (method_sel is None) or (fn_m is None) or (fn_i is None):
         raise PreventUpdate
 
@@ -208,9 +205,7 @@ def compute_value_i(method_sel, fn_m, fn_i):
 # pylint: disable=too-many-locals
 # pylint: disable=unused-argument
 def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image):
-    '''Takes in the last model and image uploaded filenames, the selected XAI method, and 
-    returns the selected XAI method'''
-
+    """Takes in the last model and image uploaded filenames, the selected XAI method, and returns the selected XAI method."""
     ctx = dash.callback_context
 
     if (ctx.triggered[0]["prop_id"] == "upload-model-img.filename") or (ctx.triggered[0]["prop_id"] == "upload-image.filename") or (not ctx.triggered):
@@ -320,6 +315,7 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image):
               dash.dependencies.Input('submit-text', 'n_clicks'),
               dash.dependencies.State('upload-text', 'value'))
 def upload_text(clicks, input_value):
+    """Takes in test text string, and print it on the dashboard."""
     if clicks is not None:
         return html.Div([
                     html.P('Input string for the model is:'),
@@ -334,6 +330,7 @@ def upload_text(clicks, input_value):
               dash.dependencies.Input('upload-model-text', 'contents'),
               dash.dependencies.State('upload-model-text', 'filename'))
 def upload_model_text(contents, filename):
+    """Takes in the model file, returns a print statement about its uploading state."""
     if contents is not None:
         try:
             if 'onnx' in filename[0]:
@@ -363,7 +360,7 @@ def upload_model_text(contents, filename):
 # and for all time.
 @cache.memoize()
 def global_store_t(method_sel, model_runner, input_text):
-
[…]
