### 2022-05-20 8dc66b8
Fix more prospector warnings
