+++ dianna/utils/tokenizers.py
@@ -1,4 +1,5 @@
-from abc import ABC, abstractmethod
+from abc import ABC
+from abc import abstractmethod
 from typing import List
 from torchtext.data import get_tokenizer
 
