### 2022-05-20 6332836
Fix linter warnings that only show up on local machine
