+++ tests/test_text_visualization.py
@@ -2,8 +2,12 @@ import os
 import shutil
 import unittest
 from pathlib import Path
+import spacy
 from dianna.visualization.text import highlight_text
 from dianna.utils.tokenizers import SpacyTokenizer
+
+
+spacy.cli.download('en_core_web_sm')
 
 
 class TextExample:
