+++ dianna/methods/rise.py
@@ -39,7 +39,7 @@ class RISE:
         self.predictions = None
         self.axis_labels = axis_labels if axis_labels is not None else []
 
-    def explain_text(self, model_or_function, input_text, labels=(0,), tokenizer=None, batch_size=100):
+    def explain_text(self, model_or_function, input_data, labels=(0,), tokenizer=None, batch_size=100):
         """Runs the RISE explainer on text.
 
            The model will be called with masked versions of the input text.
@@ -47,7 +47,7 @@ class RISE:
         Args:
             model_or_function (callable or str): The function that runs the model to be explained _or_
                                                  the path to a ONNX model on disk.
-            input_text (np.ndarray): Text to be explained
+            input_data (np.ndarray): Text to be explained
             tokenizer: Tokenizer class with tokenize and convert_tokens_to_string methods, and mask_token attribute
             labels (list(int)): Labels to be explained
             batch_size (int): Batch size to use for running the model.
@@ -56,9 +56,9 @@ class RISE:
             Explanation heatmap for each class (np.ndarray).
         """
         if tokenizer is None:
-            raise ValueError('Please provide a tokenizer to explain_text')
+            raise ValueError('Please provide a tokenizer to explain_text.')
         runner = utils.get_function(model_or_function, preprocess_function=self.preprocess_function)
-        input_tokens = np.asarray(tokenizer.tokenize(input_text))
+        input_tokens = np.asarray(tokenizer.tokenize(input_data))
         num_tokens = len(input_tokens)
         active_p_keep = self._determine_p_keep_for_text(input_tokens, runner, tokenizer) if self.p_keep is None else self.p_keep
         input_shape = (num_tokens,)
