+++ tests/test_lime.py
@@ -48,7 +48,8 @@ def test_lime_text():
     expected_word_indices = [2, 0, 3, 1]
     expected_scores = [.492, -.046, .036, -.008]
 
-    explanation = dianna.explain_text(runner, review, labels=[0], method='LIME', random_state=42)[0]
+    explanation = dianna.explain_text(runner, review, tokenizer=runner.tokenizer,
+                        labels=[0], method='LIME', random_state=42)[0]
     words = [element[0] for element in explanation]
     word_indices = [element[1] for element in explanation]
     scores = [element[2] for element in explanation]
