### 2022-06-02 7d1e606
Fix linter warning in dashboard
