### 2022-06-21 #4493 @editorialbot
[REVIEW]: DIANNA: Deep Insight And Neural Network Analysis
**Submitting author:** @elboyran (<a href="http://orcid.org/0000-0002-9834-1756">Elena Ranguelova</a>)
**Repository:** https://github.com/dianna-ai/dianna
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.2
