### 2022-06-29 401102b
Add tokenizer explain_text call in dashboard
