+++ dashboard/callbacks.py
@@ -23,7 +23,7 @@ import utilities
 from utilities import MovieReviewsModelRunner, _create_html
 import numpy as np
 import warnings
-warnings.filterwarnings('ignore') # disable warnings relateds to versions of tf
+warnings.filterwarnings('ignore')  # disable warnings relateds to versions of tf
 
 folder_on_server = "app_data"
 os.makedirs(folder_on_server, exist_ok=True)
@@ -49,11 +49,12 @@ class_name_text = ["negative", "positive"]
 
 try:
     spacy.load("en_core_web_sm")
-except Exception: # If not present, we download
+except Exception:  # If not present, we download
     spacy.cli.download("en_core_web_sm")
     spacy.load("en_core_web_sm")
 
-########################### Images page ###########################
+# ########################## Images page ###########################
+
 
 # uploading test image
 @app.callback(dash.dependencies.Output('graph_test', 'figure'),
@@ -78,9 +79,9 @@ def upload_image(contents, filename):
 
                 fig = go.Figure()
 
-                if X_test.shape[2] < 3: # it's grayscale
+                if X_test.shape[2] < 3:  # it's grayscale
 
-                    fig.add_trace(go.Heatmap(z=X_test[:,:,0], colorscale='gray', showscale=False))
+                    fig.add_trace(go.Heatmap(z=X_test[:, :, 0], colorscale='gray', showscale=False))
 
                 fig.update_layout(
                     width=300,
@@ -89,8 +90,8 @@ def upload_image(contents, filename):
                     title_x=0.5,
                     title_font_color=layouts.colors['blue1'])
 
-                fig.update_xaxes(showgrid = False, showticklabels = False, zeroline=False)
-                fig.update_yaxes(showgrid = False, showticklabels = False, zeroline=False)
+                fig.update_xaxes(showgrid=False, showticklabels=False, zeroline=False)
+                fig.update_yaxes(showgrid=False, showticklabels=False, zeroline=False)
 
                 fig.layout.paper_bgcolor = layouts.colors['blue4']
 
@@ -98,13 +99,14 @@ def upload_image(contents, filename):
 
             return utilities.blank_fig(
                 text='File format error! <br><br>Please upload only images in .jpg format.')
-        
+
         except Exception as e:
             print(e)
             return utilities.blank_fig(
                     text='There was an error processing this file.')
     else:
         return utilities.blank_fig()
+
 
 # uploading model for image
 @app.callback(dash.dependencies.Output('output-model-img-upload', 'children'),
@@ -134,6 +136,7 @@ def upload_model_img(contents, filename):
     else:
         raise PreventUpdate
 
+
 # perform expensive computations in this "global store"
 # these computations are cached in a globally available
 # redis memory store which is available across processes
@@ -147,32 +150,33 @@ def global_store_i(method_sel, model_path, image_test):
             model_path, image_test, method=method_sel,
             labels=list(range(2)),
             n_masks=5000, feature_res=8, p_keep=.1,
-            axis_labels=('height','width','channels'))
+            axis_labels=('height', 'width', 'channels'))
 
     elif method_sel == "KernelSHAP":
         relevances = dianna.explain_image(
             model_path, image_test,
             method=method_sel, nsamples=1000,
             background=0, n_segments=200, sigma=0,
-            axis_labels=('height','width','channels'))
+            axis_labels=('height', 'width', 'channels'))
 
     else:
         relevances = dianna.explain_image(
             model_path, image_test * 256, 'LIME',
-            axis_labels=('height','width','channels'),
+            axis_labels=('height', 'width', 'channels'),
             random_state=2,
             labels=list(range(2)),
             preprocess_function=utilities.preprocess_function)
 
     return relevances
 
+
 # signaling
[…]
