### 2022-06-29 92e05df
refactor generate masks for images to static method
