### 2022-06-29 3ca342f
add distance explanation function, refs #279
