+++ dianna/__init__.py
@@ -24,6 +24,7 @@ See https://github.com/dianna-ai/dianna
 import importlib
 import logging
 from . import utils
+from .methods.distance import DistanceExplainer
 
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
@@ -74,6 +75,12 @@ def explain_text(model_or_function, input_data, method, labels=(1,), **kwargs):
     explainer = _get_explainer(method, kwargs)
     explain_text_kwargs = utils.get_kwargs_applicable_to_function(explainer.explain_text, kwargs)
     return explainer.explain_text(model_or_function, input_data, labels, **explain_text_kwargs)
+
+def explain_image_distance(model_or_function, input_data, embedded_reference, **kwargs):
+    method_kwargs = utils.get_kwargs_applicable_to_function(DistanceExplainer.__init__, kwargs)
+    explainer = DistanceExplainer(**method_kwargs)
+    explain_distance_kwargs = utils.get_kwargs_applicable_to_function(explainer.explain_image_distance, kwargs)
+    return explainer.explain_image_distance(model_or_function, input_data, embedded_reference, **explain_distance_kwargs)
 
 
 def _get_explainer(method, kwargs):
