+++ dianna/methods/distance.py
@@ -47,7 +47,6 @@ class DistanceExplainer:
         img_shape = input_data.shape[1:3]
         # Expose masks for to make user inspection possible
         self.masks = generate_masks_for_images(img_shape, active_p_keep, self.n_masks, self.feature_res)
-        
         # Make sure multiplication is being done for correct axes
         masked = input_data * self.masks
 
@@ -58,7 +57,7 @@ class DistanceExplainer:
 
         reference_pred = embedded_reference
         distances = pairwise_distances(self.predictions, reference_pred, metric='cosine') / 2
-        lowest_distances_indices = np.argsort(distances, axis=0)[:int(len(self.predictions) * self.keep_lowest_distance_masks_fraction)]
+        lowest_distances_indices = np.argsort(distances, axis=0)[:int(len(self.predictions) * self.p_keep_lowest_distances)]
 
         mask_weights = np.exp(-distances[lowest_distances_indices])
         lowest_distances_masks = self.masks[lowest_distances_indices]
