### 2022-07-08 abb99d2
Use list comprehension to avoid looping over local_explanation twice
Co-authored-by: Christiaan Meijer <c.meijer@esciencecenter.nl>
