### 2022-07-08 55549d6
Remove unnecessary _find_token_indices
