### 2022-07-08 3bf8688
Remove removed parameter from docstring
