### 2022-07-08 0ebf4c2
Rename input_data to input_text in all text related methods.
