+++ dianna/methods/rise.py
@@ -39,7 +39,7 @@ class RISE:
         self.predictions = None
         self.axis_labels = axis_labels if axis_labels is not None else []
 
-    def explain_text(self, model_or_function, input_data, labels=(0,), tokenizer=None, batch_size=100):
+    def explain_text(self, model_or_function, input_text, labels=(0,), tokenizer=None, batch_size=100):
         """Runs the RISE explainer on text.
 
            The model will be called with masked versions of the input text.
@@ -47,7 +47,7 @@ class RISE:
         Args:
             model_or_function (callable or str): The function that runs the model to be explained _or_
                                                  the path to a ONNX model on disk.
-            input_data (np.ndarray): Text to be explained
+            input_text (np.ndarray): Text to be explained
             tokenizer: Tokenizer class with tokenize and convert_tokens_to_string methods, and mask_token attribute
             labels (list(int)): Labels to be explained
             batch_size (int): Batch size to use for running the model.
@@ -58,7 +58,7 @@ class RISE:
         if tokenizer is None:
             raise ValueError('Please provide a tokenizer to explain_text.')
         runner = utils.get_function(model_or_function, preprocess_function=self.preprocess_function)
-        input_tokens = np.asarray(tokenizer.tokenize(input_data))
+        input_tokens = np.asarray(tokenizer.tokenize(input_text))
         num_tokens = len(input_tokens)
         active_p_keep = self._determine_p_keep_for_text(input_tokens, runner, tokenizer) if self.p_keep is None else self.p_keep
         input_shape = (num_tokens,)
@@ -68,22 +68,22 @@ class RISE:
         saliencies = self._get_saliencies(runner, masked_sentences, num_tokens, batch_size, active_p_keep)
         return self._reshape_result(input_tokens, labels, saliencies)
 
-    def _determine_p_keep_for_text(self, input_data, runner, tokenizer, n_masks=100):
+    def _determine_p_keep_for_text(self, input_text, runner, tokenizer, n_masks=100):
         """See n_mask default value https://github.com/dianna-ai/dianna/issues/24#issuecomment-1000152233."""
         p_keeps = np.arange(0.1, 1.0, 0.1)
         stds = []
         for p_keep in p_keeps:
-            std = self._calculate_mean_class_std_for_text(p_keep, runner, input_data, tokenizer, n_masks=n_masks)
+            std = self._calculate_mean_class_std_for_text(p_keep, runner, input_text, tokenizer, n_masks=n_masks)
             stds += [std]
         best_i = np.argmax(stds)
         best_p_keep = p_keeps[best_i]
         print(f'Rise parameter p_keep was automatically determined at {best_p_keep}')
         return best_p_keep
 
-    def _calculate_mean_class_std_for_text(self, p_keep, runner, input_data, tokenizer, n_masks):
+    def _calculate_mean_class_std_for_text(self, p_keep, runner, input_text, tokenizer, n_masks):
         batch_size = 50
-        masks = self._generate_masks_for_text(input_data.shape, p_keep, n_masks)
-        masked = self._create_masked_sentences(input_data, masks, tokenizer)
+        masks = self._generate_masks_for_text(input_text.shape, p_keep, n_masks)
+        masked = self._create_masked_sentences(input_text, masks, tokenizer)
         predictions = []
         for i in range(0, n_masks, batch_size):
             current_input = masked[i:i + batch_size]
