### 2022-07-13 c2b661e
add tests for distance explainer functionality
