### 2022-07-14 f89ef6b
add exact result test for distance
