### 2022-07-14 95c6f65
refactor distance
