+++ dianna/methods/distance.py
@@ -20,7 +20,7 @@ from sklearn.metrics import pairwise_distances
 
 class DistanceExplainer:
     # axis labels required to be present in input image data
-    required_labels = ('channels', )
+    required_labels = ('channels',)
 
     def __init__(self, n_masks=1000, feature_res=8, p_keep=.5,  # pylint: disable=too-many-arguments
                  p_keep_lowest_distances=.2, axis_labels=None, batch_size=10, preprocess_function=None):
@@ -35,12 +35,9 @@ class DistanceExplainer:
         self.batch_size = batch_size
 
     def explain_image_distance(self, model_or_function, input_data, embedded_reference, **explain_distance_kwargs):
-        input_data = utils.to_xarray(input_data, self.axis_labels, DistanceExplainer.required_labels)
-        # add batch axis as first axis
-        input_data = input_data.expand_dims('batch', 0)
-        input_data, full_preprocess_function = self._prepare_image_data(input_data)
+        full_preprocess_function, input_data = self._prepare_input_data(input_data)
         runner = utils.get_function(model_or_function, preprocess_function=full_preprocess_function)
-        
+
         active_p_keep = 0.5 if self.p_keep is None else self.p_keep  # Could autotune here (See #319)
 
         # data shape without batch axis and channel axis
@@ -50,26 +47,42 @@ class DistanceExplainer:
         # Make sure multiplication is being done for correct axes
         masked = input_data * self.masks
 
-        batch_predictions = []
-        for i in tqdm(range(0, self.n_masks, self.batch_size), desc='Explaining'):
-            batch_predictions.append(runner(masked[i:i + self.batch_size]))
+        batch_predictions = [runner(masked[i:i + self.batch_size])
+                             for i in tqdm(range(0, self.n_masks, self.batch_size), desc='Explaining')]
         self.predictions = np.concatenate(batch_predictions)
 
-        reference_pred = embedded_reference
-        distances = pairwise_distances(self.predictions, reference_pred, metric='cosine') / 2  # divide by 2 to have [0.1] output range
-        lowest_distances_indices = np.argsort(distances, axis=0)[:int(len(self.predictions) * self.p_keep_lowest_distances)]
+        lowest_distances_masks, mask_weights = self._get_lowest_distance_masks_and_weights(embedded_reference,
+                                                                                           self.predictions, self.masks,
+                                                                                           self.p_keep_lowest_distances)
 
-        mask_weights = np.exp(-distances[lowest_distances_indices])
-        lowest_distances_masks = self.masks[lowest_distances_indices]
-
-        sal = mask_weights.T.dot(lowest_distances_masks.reshape(len(lowest_distances_masks), -1)).reshape(-1, *img_shape)
+        sal = mask_weights.T.dot(lowest_distances_masks.reshape(len(lowest_distances_masks), -1)).reshape(-1,
+                                                                                                          *img_shape)
 
         normalization = mask_weights.sum()
 
-        sal = sal / normalization
-        return sal
+        return sal / normalization
 
-    
+    @staticmethod
+    def _get_lowest_distance_masks_and_weights(embedded_reference, predictions, masks, p_keep_lowest_distances):
+        distances = pairwise_distances(predictions, embedded_reference,
+                                       metric='cosine') / 2  # divide by 2 to have [0.1] output range
+        lowest_distances_indices = np.argsort(distances, axis=0)[
+                                   :int(len(predictions) * p_keep_lowest_distances)]
+        mask_weights = np.exp(-distances[lowest_distances_indices])
+        lowest_distances_masks = masks[lowest_distances_indices]
+        return lowest_distances_masks, mask_weights
+
+    def _prepare_input_data(self, input_data):
[…]
