### 2022-07-28 #4493 @Athene-ai
**More comments**:

_Installation: Does installation proceed as outlined in the documentation?_ : The reviewer suggests improving this part of the paper to explain in more details the installation procedure to newbie

_Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?_ : The reviewer suggests to revise the way of writing this paper

_Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?_ : This part of the paper should be implemented

_Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution._: This part should be better explained

_Functionality_: This part should be better implemented and explained
