### 2022-08-02 #4493 @elboyran
Dear @sara-02 and @Athene-ai thank you for your reviews.

@taless474 I see that some check-boxes are not checked from both of the reviewers and @Athene-ai gave some comments, though very brief and generic and I cannot see comments from @sara-02.

How shall I proceed? May I contact the reviewers for more information/suggestions than given?
For example I am not sure how to approach "The reviewer suggests to revise the way of writing this paper", it is not clear what should be changed.

Should all check-boxes be ticked for the paper to proceed?
