### 2022-08-02 #4493 @sara-02
@elboyran For me, the paper content is acceptable. I have an issue that my automated tests failed. I will create an issue to follow up. The installation setup worked fine but to test the functionality, I need to have a model at hand. Is it possible to share a small MNIST based model or some pretrained model on hugging face that can be used to independently test the functionalities of your system?
