### 2022-08-02 #4493 @elboyran
> @elboyran For me, the paper content is acceptable. I have an issue that my automated tests failed. I will create an issue to follow up. The installation setup worked fine but to test the functionality, I need to have a model at hand. Is it possible to share a small MNIST based model or some pretrained model on hugging face that can be used to independently test the functionalities of your system?

Dear @sara-02 , the DIANNA README lists all datasets and models we have used in our project. Please, have a look at
https://github.com/dianna-ai/dianna#images-1 for a list of links to ONNX models trained on image datasets, including a 'binary MNIST' model.
