### 2022-08-03 31df975
fix small typos in readme
