### 2022-09-01 c024cb5
expose SimpleModelRunner to user
I think a user is not able to import the SimpleModelRunner at the moment. This commit exposes it to the user so they can reuse it.
