+++ dianna/utils/__init__.py
@@ -5,3 +5,4 @@ from .misc import move_axis
 from .misc import onnx_model_node_loader
 from .misc import to_xarray
 from .tokenizers import SpacyTokenizer
+from .onnx_runner import SimpleModelRunner
