### 2022-09-06 0bfa98d
First attempt at splitting RISE into separate classes for text and images
