+++ dianna/methods/rise.py
@@ -31,10 +31,9 @@ class RISEText:
         self.p_keep = p_keep
         self.preprocess_function = preprocess_function
         self.masks = None
-        self.mask_string = ""
         self.predictions = None
 
-    def explain(self, model_or_function, input_text, tokenizer, labels=(0,), batch_size=100):
+    def explain(self, model_or_function, input_text, labels=(0,), tokenizer=None, batch_size=100):
         """Runs the RISE explainer on text.
 
            The model will be called with masked versions of the input text.
@@ -43,41 +42,43 @@ class RISEText:
             model_or_function (callable or str): The function that runs the model to be explained _or_
                                                  the path to a ONNX model on disk.
             input_text (np.ndarray): Text to be explained
-            Tokenizer class with tokenize and convert_tokens_to_string methods, and mask_token attribute
+            tokenizer: Tokenizer class with tokenize and convert_tokens_to_string methods, and mask_token attribute
             labels (list(int)): Labels to be explained
             batch_size (int): Batch size to use for running the model.
 
         Returns:
             Explanation heatmap for each class (np.ndarray).
         """
+        if tokenizer is None:
+            raise ValueError('Please provide a tokenizer to explain_text.')
+
         runner = utils.get_function(model_or_function, preprocess_function=self.preprocess_function)
-        self.mask_string = tokenizer.mask_token
         input_tokens = np.asarray(tokenizer.tokenize(input_text))
-        text_length = len(input_tokens)
-        active_p_keep = self._determine_p_keep(input_tokens, runner) if self.p_keep is None else self.p_keep
-        input_shape = (text_length,)
+        num_tokens = len(input_tokens)
+        active_p_keep = self._determine_p_keep(input_tokens, runner, tokenizer) if self.p_keep is None else self.p_keep
+        input_shape = (num_tokens, )
         self.masks = self._generate_masks(input_shape, active_p_keep,
                                           self.n_masks)  # Expose masks for to make user inspection possible
-        sentences = self._create_masked_sentences(input_tokens, self.masks)
-        saliencies = self._get_saliencies(runner, sentences, text_length, batch_size, active_p_keep)
+        masked_sentences = self._create_masked_sentences(input_tokens, self.masks, tokenizer)
+        saliencies = self._get_saliencies(runner, masked_sentences, num_tokens, batch_size, active_p_keep)
         return self._reshape_result(input_tokens, labels, saliencies)
 
-    def _determine_p_keep(self, input_data, runner, n_masks=100):
+    def _determine_p_keep(self, input_text, runner, tokenizer, n_masks=100):
         """See n_mask default value https://github.com/dianna-ai/dianna/issues/24#issuecomment-1000152233."""
         p_keeps = np.arange(0.1, 1.0, 0.1)
         stds = []
         for p_keep in p_keeps:
-            std = self._calculate_mean_class_std(p_keep, runner, input_data, n_masks=n_masks)
+            std = self._calculate_mean_class_std(p_keep, runner, input_text, tokenizer, n_masks=n_masks)
             stds += [std]
         best_i = np.argmax(stds)
         best_p_keep = p_keeps[best_i]
         print(f'Rise parameter p_keep was automatically determined at {best_p_keep}')
         return best_p_keep
 
-    def _calculate_mean_class_std(self, p_keep, runner, input_data, n_masks):
+    def _calculate_mean_class_std(self, p_keep, runner, input_text, tokenizer, n_masks):
         batch_size = 50
-        masks = self._generate_masks(input_data.shape, p_keep, n_masks)
-        masked = self._create_masked_sentences(input_data, masks)
[…]
