### 2022-09-07 0aff6c3
Fix text visualization in dashboard
