### 2022-09-07 661a73a
Rename KernelSHAP class to KernelSHAPImage
