### 2022-09-07 3873efa
Split LIME into separate class for images and text
