### 2022-09-07 f5c3dd5
Always use new way of importing method class in init.py, i.e. including modality
