+++ dianna/__init__.py
@@ -52,14 +52,8 @@ def explain_image(model_or_function, input_data, method, labels=(1,), **kwargs):
         # To avoid Access Violation on Windows with SHAP:
         from onnx_tf.backend import prepare  # pylint: disable=import-outside-toplevel,unused-import
     explainer = _get_explainer(method, kwargs, modality="Image")
-    # the name of the explainer function is different for RISE while it is the only method split
-    # into separate classes per data modality
-    if method.lower() == "rise":
-        explainer_function = explainer.explain
-    else:
-        explainer_function = explainer.explain_image
-    explain_image_kwargs = utils.get_kwargs_applicable_to_function(explainer_function, kwargs)
-    return explainer_function(model_or_function, input_data, labels, **explain_image_kwargs)
+    explain_image_kwargs = utils.get_kwargs_applicable_to_function(explainer.explain, kwargs)
+    return explainer.explain(model_or_function, input_data, labels, **explain_image_kwargs)
 
 
 def explain_text(model_or_function, input_text, tokenizer, method, labels=(1,), **kwargs):
@@ -79,14 +73,8 @@ def explain_text(model_or_function, input_text, tokenizer, method, labels=(1,), 
 
     """
     explainer = _get_explainer(method, kwargs, modality="Text")
-    # the name of the explainer function is different for RISE while it is the only method split
-    # into separate classes per data modality
-    if method.lower() == "rise":
-        explainer_function = explainer.explain
-    else:
-        explainer_function = explainer.explain_text
-    explain_text_kwargs = utils.get_kwargs_applicable_to_function(explainer_function, kwargs)
-    return explainer_function(
+    explain_text_kwargs = utils.get_kwargs_applicable_to_function(explainer.explain, kwargs)
+    return explainer.explain(
         model_or_function=model_or_function,
         input_text=input_text,
         labels=labels,
@@ -94,14 +82,14 @@ def explain_text(model_or_function, input_text, tokenizer, method, labels=(1,), 
         **explain_text_kwargs)
 
 
-def _get_explainer(method, kwargs, modality=None):
-    method_submodule = importlib.import_module(f'dianna.methods.{method.lower()}')
-    # The way of getting the explainer is different for RISE while RISE is the only method
-    # that is split into separate classes per data modality
-    if method.lower() == "rise":
-        assert modality is not None, "Data modality is mandatory when method is set to RISE"
+def _get_explainer(method, kwargs, modality):
+    try:
+        method_submodule = importlib.import_module(f'dianna.methods.{method.lower()}')
+    except ImportError:
+        raise Exception(f"Method {method} does not exist")
+    try:
         method_class = getattr(method_submodule, f"{method}{modality}")
-    else:
-        method_class = getattr(method_submodule, method)
+    except AttributeError:
+        raise Exception(f"Data modality {modality} is not available for method {method}")
     method_kwargs = utils.get_kwargs_applicable_to_function(method_class.__init__, kwargs)
     return method_class(**method_kwargs)
