+++ dianna/__init__.py
@@ -85,11 +85,11 @@ def explain_text(model_or_function, input_text, tokenizer, method, labels=(1,), 
 def _get_explainer(method, kwargs, modality):
     try:
         method_submodule = importlib.import_module(f'dianna.methods.{method.lower()}')
-    except ImportError:
-        raise Exception(f"Method {method} does not exist")
+    except ImportError as err:
+        raise Exception(f"Method {method} does not exist") from err
     try:
         method_class = getattr(method_submodule, f"{method}{modality}")
-    except AttributeError:
-        raise Exception(f"Data modality {modality} is not available for method {method}")
+    except AttributeError as err:
+        raise Exception(f"Data modality {modality} is not available for method {method}") from err
     method_kwargs = utils.get_kwargs_applicable_to_function(method_class.__init__, kwargs)
     return method_class(**method_kwargs)
