### 2022-09-12 #4493 @Athene-ai
Sure @elboyran

Here there are more comments

_Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? : The reviewer suggests to revise the way of writing this paper_  With this expression I would like to say that the author should better write the paper
