### 2022-09-12 #4493 @Athene-ai
English language

On Monday, 12 September 2022, Elena Ranguelova ***@***.***>
wrote:

> Sure @elboyran <https://github.com/elboyran>
>
> Here there are more comments
>
> *Quality of writing: Is the paper well written (i.e., it does not require
> editing for structure, language, or writing quality)? : The reviewer
> suggests to revise the way of writing this paper* --> With this
> expression I would like to say that the author should better write the paper
>
> Thank you, but I am still unclear in what respect it should be better? The
> (English) language? The length? The structure? Anything else?
>
