### 2022-09-14 8bd6324
add neutral value to return values of explain distance
