### 2022-09-19 #4493 @sara-02
> > > @elboyran For me, the paper content is acceptable. I have an issue that my automated tests failed. I will create an issue to follow up. The installation setup worked fine but to test the functionality, I need to have a model at hand. Is it possible to share a small MNIST based model or some pretrained model on hugging face that can be used to independently test the functionalities of your system?
> >
> >
> > Dear @sara-02 , the DIANNA README lists all datasets and models we have used in our project. Please, have a look at https://github.com/dianna-ai/dianna#images-1 for a list of links to ONNX models trained on image datasets, including a 'binary MNIST' model.
> > Apart from Zenodo, all models used in our tutorials are also available in the DIANNA github repo at [tutorials/models](https://github.com/dianna-ai/dianna/tree/main/tutorials/models).
>
> Dear @sara-02, did it work for you now? Are more boxes in your review to be ticked?

Last I was trying I was facing issues running the tests, I have updated by OS since then let me try again and revert asap.
