+++ dianna/utils/__init__.py
@@ -4,5 +4,5 @@ from .misc import get_kwargs_applicable_to_function
 from .misc import move_axis
 from .misc import onnx_model_node_loader
 from .misc import to_xarray
-from .tokenizers import SpacyTokenizer
 from .onnx_runner import SimpleModelRunner
+from .tokenizers import SpacyTokenizer
