+++ setup.cfg
@@ -26,7 +26,7 @@ name = dianna
 project_urls =
     Bug Tracker = https://github.com/dianna-ai/dianna/issues
 url = https://github.com/dianna-ai/dianna
-version = 0.4.1
+version = 0.5.0
 
 [options]
 zip_safe = False
