### 2022-09-20 dianna-ai/dianna#354 @elboyran
JOSS: address reviews
Address the [JOSS reviews](https://github.com/openjournals/joss-reviews/issues/4493)
