### 2022-09-20 b9d1f91
Address reviewer suggestion about Used by
I tried to improve the English of this section, as suggested in our JOSS paper review. Feel free to edit my changes / add others.
