### 2022-09-20 dianna-ai/dianna#354 @gcroci2
I opened PR #355
