### 2022-09-22 694fa42
Update standup-checklist.md
