### 2022-09-27 66a3704
Update standup-checklist.md
