### 2022-10-04 c519756
Update standup-checklist.md
