### 2022-10-04 cd1c484
allow png and jpeg file extensions for image upload
