### 2022-10-04 afbee38
run dashboard with python script
