### 2022-10-04 066cdb8
add example notebook, pretrained model, sample data
