+++ dashboard/dashboard.py
@@ -2,18 +2,17 @@
 import dash
 from dash import html, dcc
 # Custom libraries
-from dianna import utils
+from dianna import utils  # pylint: disable=unused-import
 import layouts
 from layouts import images_page, text_page
 import utilities
-import callbacks
 from callbacks import app
 # Others
 import warnings
 warnings.filterwarnings('ignore') # disable warnings relateds to versions of tf
-
 # debug
 from importlib import reload
+
 reload(layouts)
 reload(utilities)
 
@@ -25,10 +24,10 @@ app.layout = html.Div([
 @app.callback(dash.dependencies.Output('page-content', 'children'),
               [dash.dependencies.Input('url', 'pathname')])
 def display_page(pathname):
+    """load content of dashboard"""
     if pathname == '/apps/text':
-         return text_page
-    else:
-         return images_page # home page
+        return text_page
+    return images_page # home page
 
 
 if __name__ == '__main__':
