### 2022-10-05 785a058
add support for 2 ranges wrt distance weights
