### 2022-10-05 f5a1ae4
add support for PIL images
