### 2022-10-06 3260531
upload imagenet classnames, and show uploaded rgb fig
