### 2022-10-18 ad09a0b
Update README.md
added more model formats in Getting Started link
