### 2022-10-18 9c710b4
Add heatmaps to compare RISE tests output to earlier output
