### 2022-10-18 9da1e67
Update README.md
addresses issue #336
