### 2022-10-19 f08fbde
Use absolute tolerance of 1e-5 when comparing test output to expected output
