### 2022-10-19 475c5f0
Avoid use of general exception
