### 2022-10-19 5c77fa6
Add docstring to custom exception
