+++ dianna/__init__.py
@@ -34,6 +34,7 @@ __version__ = "0.5.0"
 
 
 class XAIMethodException(Exception):
+    """DIANNA exception used when the user asks for a non-existent method or data modality"""
     pass
 
 
