### 2022-10-19 dianna-ai/dianna#354 @egpbos
So

1. an installation instruction should be added (`pip install dianna`, basically? the reviewer says "for newbies", so maybe they also want to see that you first must install python, pip and only then install dianna?) and
2. a "functionality" explanation should be added, so I guess a basic usage example?
3. Depending on whether sara-02 can run the tests now, something may need to be added there as well?
