### 2022-10-24 381cae7
first (hacky) attempt to implement imagenet, including option to show top prediction
