+++ dashboard/callbacks.py
@@ -17,12 +17,13 @@ from PIL import Image
 from html2image import Html2Image
 import dianna
 from dianna.utils.tokenizers import SpacyTokenizer
+from keras import utils as keras_utils
 import spacy
 import os
 import base64
 import layouts
 import utilities
-from utilities import MovieReviewsModelRunner, _create_html, imagenet_class_name
+from utilities import MovieReviewsModelRunner, _create_html, imagenet_class_name, Model_imagenet
 import numpy as np
 import warnings
 warnings.filterwarnings('ignore')  # disable warnings relateds to versions of tf
@@ -46,11 +47,10 @@ cache = Cache(app.server, config={
 })
 cache.clear()
 
-# global variables
+# global variables # replace by generic label loader
 class_name_mnist = ['digit 0', 'digit 1']
 class_name_text = ["negative", "positive"]
 class_names_imagenet = [imagenet_class_name(idx) for idx in range(1000)]
-class_names_imagenet = class_names_imagenet[:4]
 
 try:
     spacy.load("en_core_web_sm")
@@ -79,14 +79,14 @@ def upload_image(contents, filename):
 
                 data_path = os.path.join(folder_on_server, filename[0])
 
-                X_test = utilities.open_image(data_path)
-                
+                X_test, img = utilities.open_image(data_path)
 
                 if X_test.shape[2] < 3:  # it's grayscale
                     fig = go.Figure()
                     fig.add_trace(go.Heatmap(z=X_test[:, :, 0], colorscale='gray', showscale=False))
                 else: # it's multicolor
-                    fig = px.imshow(X_test)
+                    X_test, img = utilities.open_image(data_path)
+                    fig = px.imshow(img)
 
                 fig.update_layout(
                     width=300,
@@ -156,7 +156,6 @@ def global_store_i(method_sel, model_path, image_test, labels=list(range(2))):
             labels=labels,
             n_masks=5000, feature_res=8, p_keep=.1,
             axis_labels=('height', 'width', 'channels'))
-
     elif method_sel == "KernelSHAP":
         relevances = dianna.explain_image(
             model_path, image_test,
@@ -171,7 +170,6 @@ def global_store_i(method_sel, model_path, image_test, labels=list(range(2))):
             random_state=2,
             labels=labels,
             preprocess_function=utilities.preprocess_function)
-
     return relevances
 
 
@@ -190,7 +188,7 @@ def compute_value_i(method_sel, fn_m, fn_i):
     for m in method_sel:
         # compute value and send a signal when done
         data_path = os.path.join(folder_on_server, fn_i[0])
-        image_test = utilities.open_image(data_path)
+        image_test, _ = utilities.open_image(data_path)
 
         model_path = os.path.join(folder_on_server, fn_m[0])
 
@@ -211,10 +209,11 @@ def compute_value_i(method_sel, fn_m, fn_i):
     dash.dependencies.Input("signal_image", "data"),
     dash.dependencies.Input("upload-model-img", "filename"),
     dash.dependencies.Input("upload-image", "filename"),
+    dash.dependencies.Input("show_top", "value"),
 )
 # pylint: disable=too-many-locals
 # pylint: disable=unused-argument
-def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image):
+def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_top=2):
     """Takes in the last model and image uploaded filenames, the selected XAI method, and returns the selected XAI method."""
     ctx = dash.callback_context
 
@@ -229,7 +228,7 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image):
[…]
