### 2022-10-24 f66e58d
load colour images and do prep for resnet model
