### 2022-10-25 531d848
no more mask weighting for distance
