### 2022-10-25 ea7ad61
decreased taxonomy image size
It is hard to come up with something that looks great on all screen sizes. I think this (60%) version looks better on desktops/laptops which have in general a larger screen than other kind of devices. For smaller devices like a phone, the 100% version was maybe better but I'm guessing most of our audience will browse with a laptop/desktop.
