### 2022-10-25 c6ab3b6
Change XaiMethodException to ValueError
Seems a better balance between specificity and keeping things clean and expected.
