### 2022-10-25 4437ee9
Make torchtext optional by only requiring it when user explicitly imports a tokenizer. Add dianna[text] variant that installs torchtext
