+++ dianna/utils/__init__.py
@@ -5,4 +5,3 @@ from .misc import move_axis
 from .misc import onnx_model_node_loader
 from .misc import to_xarray
 from .onnx_runner import SimpleModelRunner
-from .tokenizers import SpacyTokenizer
