+++ setup.cfg
@@ -88,6 +88,8 @@ notebooks =
     torchtext
     torchvision
     ipywidgets
+text =
+    torchtext
 
 [options.packages.find]
 include = dianna, dianna.*
