+++ dianna/utils/tokenizers.py
@@ -5,9 +5,9 @@ from typing import List
 
 try:
     from torchtext.data import get_tokenizer
-except ImportError:
+except ImportError as err:
     raise ImportError("Failed to import torchtext, please install manually or reinstall dianna with "
-                      "text support: `pip install dianna[text]`")
+                      "text support: `pip install dianna[text]`") from err
 
 
 class Tokenizer(ABC):
