### 2022-10-25 5b987a8
Add missing spacy dependency to text variant
