+++ setup.cfg
@@ -90,6 +90,7 @@ notebooks =
     ipywidgets
 text =
     torchtext
+    spacy
 
 [options.packages.find]
 include = dianna, dianna.*
