### 2022-10-26 #4493 @sara-02
@Athene-ai I was able to pip install and successfully import in python 3.8 Suggestion for improving 1st time trying folks, to train the MNIST model, https://github.com/dianna-ai/dianna-exploration/tree/main/example_data/model_generation/MNIST can the dataset be loaded from skelearn dataset itself instead of from a path?
