### 2022-10-26 #4493 @sara-02
In line with the above ipynb and the documentations readthedocs, (maybe I missing it) but can there be one example that runs end-to-end like the same notebook has model training as well as explanation part instead of training in one and assuming loading in the read the docs. This can help give the newcomers a sense of how the workflow is end2end.
