### 2022-10-26 #4493 @sara-02
@elboyran I was able to follow the testing setup https://dianna.readthedocs.io/en/latest/developer_info.html mentioned here. My tests with tox (conda python3.8 env) worked as given in the screenshot.
However running the pytest -v command from the same test folder gives an error

![image](https://user-images.githubusercontent.com/7047079/197971958-dffc9646-2a3b-4c8c-b4d7-6c018649daf9.png)
