### 2022-10-28 5aeabd6
remove initial hacky attempt to get Imagenet to work
