### 2022-10-28 b7679bf
prep colour image for Imagenet (with RISE)
