### 2022-10-31 #4493 @cwmeijer
Dear @sara-02,

I'm one of the co-authors of the paper. Thank you both for the suggestions you made above. I created issues to address them (soon) and improve our repository.

I see one unticked box left "Functionality: Have the functional claims of the software been confirmed?" and I'm not clear how to interpret this. Did you read about specific functionality that you couldn't find in our software or were you just not able to try everything out yet?

Thanks a lot,
Christiaan
