### 2022-10-31 2d1b1f3
make axis labels into argument, add specifications for resnet relevances and plots
