### 2022-10-31 ed25776
correct return norm_img_data
