### 2022-11-01 19234ae
update plotting to make rise work for resnet and mnist
