### 2022-11-01 dianna-ai/dianna#354 @cwmeijer
Outdated and duplicate of #367 .
