### 2022-11-02 5df55ac
log statistics to explainer object instead of print
