### 2022-11-02 ccb2f11
add selector buttons for XAI method settings
