+++ dashboard/layouts.py
@@ -244,38 +244,189 @@ def get_uploads_images():
                 'textAlign': 'center',
                 'align-items': 'center'
             }),
-
+        
         # XAI methods col (9-col)
         html.Div([
 
             # XAI methods selection row
             html.Div([
-                dcc.Dropdown(id = 'method_sel_img',
-                    options = [{'label': 'RISE', 'value': 'RISE'},
-                            {'label': 'KernelSHAP', 'value': 'KernelSHAP'},
-                            {'label': 'LIME', 'value': 'LIME'}],
-                    placeholder = "Select one/more XAI methods",
-                    value=[""],
-                    multi = True,
-                    style={
-                            'margin-left': '155px',
-                            'margin-top': '20px',
-                            'width': '60%',
-                            'color' : colors['blue1']
-                        }
+                html.Div([
+                    dcc.Markdown(children='**Select one/more XAI methods**',
+                        style = {'margin-top': '20px', 'textAlign' : 'center', 'color' : colors['blue1']}
+                        ),
+                    dcc.Dropdown(id = 'method_sel_img',
+                        options = [{'label': 'RISE', 'value': 'RISE'},
+                                {'label': 'KernelSHAP', 'value': 'KernelSHAP'},
+                                {'label': 'LIME', 'value': 'LIME'}],
+                        placeholder = "Select methods",
+                        value=[""],
+                        multi = True,
+                        style={
+                                'margin-left': '0px',
+                                'margin-top': '10px',
+                                'width': '100%',
+                                'color' : colors['blue1']
+                            }
+                    )
+                ], className = 'six columns'
                 ),
-                dcc.Input(id = 'show_top',
-                    placeholder = "Show top #",
-                    type="number",
-                    value=2,
-                    debounce=True,
-                    style={
-                            'margin-right': '320px',
-                            'margin-top': '20px',
-                            'width': '15%',
-                            'color' : colors['blue1']
-                        }
-                )
+                html.Div([
+                    dcc.Markdown(children='**Number of top results to show**',
+                        style = {'margin-top': '20px', 'textAlign' : 'center', 'color' : colors['blue1']}
+                        ),
+                    dcc.Input(id = 'show_top',
+                            type="text",
+                            value=2,
+                            style={
+                                    'margin-right': '10px',
+                                    'margin-top': '2px',
+                                    'width': '100%',
+                                    'color' : colors['blue1']
+                                }
+                        )
+                ], className = 'three columns'
+                ),
+            ],
+            className = 'row', style = {'padding-bottom' : '3%'}
+            ),
+            # Settings bar
+            html.Div([
+                html.Div([
+                    html.H6(children='XAI method specific settings', style={'font-weight': 'bold'}),
+                ], className='nine columns'
+                ),
+            ],
+            className = 'row'
+            ),
+            # XAI method settings buttons
+            html.Div([
+                html.Div([
+                    dcc.Markdown(children='**Rise**',
+                        style = {'margin-top': '20px', 'textAlign' : 'center', 'color' : colors['blue1']}
+                        ),
+                    dcc.Markdown(children='N masks',
+                        style = {'margin-left': '5px', 'margin-top': '0px', 'textAlign' : 'left', 'color' : colors['blue1']}
+                        ),
+                    dcc.Input(id = 'n_masks',
+                        placeholder = "N masks",
+                        type="number",
+                        value=1000,
+                        style={
+                                'margin-right': '20px',
+                                'margin-top': '0px',
+                                'width': '100%',
+                                'color' : colors['blue1']
+                            }
+                    ),
+                    dcc.Markdown(children='Feature res',
+                        style = {'margin-left': '5px', 'margin-top': '5px', 'textAlign' : 'left', 'color' : colors['blue1']}
+                        ),
+                    dcc.Input(id = 'feature_res',
[…]
