### 2022-11-02 47824e0
make dashboard inputs interactive
