### 2022-11-04 b163c0b
fix input type of show_top
