### 2022-11-04 8e7956e
make code more generic, fix mirroring heatmaps, fix show top exceeding number of class names, change layout predicted class
