+++ dashboard/callbacks.py
@@ -81,13 +81,8 @@ def upload_image(contents, filename):
 
                 data_path = os.path.join(folder_on_server, filename[0])
 
-                X_test, img = utilities.open_image(data_path)
-
-                if X_test.shape[2] < 3:  # it's grayscale
-                    fig = go.Figure()
-                    fig.add_trace(go.Heatmap(z=X_test[:, :, 0], colorscale='gray', showscale=False))
-                else: # it's multicolor
-                    fig = px.imshow(img)
+                _, img = utilities.open_image(data_path)
+                fig = px.imshow(img)
 
                 fig.update_layout(
                     width=300,
@@ -239,7 +234,7 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_t
     if (fn_m and fn_i) is not None:
 
         data_path = os.path.join(folder_on_server, fn_i[0])
-        X_test, img = utilities.open_image(data_path)
+        X_test, _ = utilities.open_image(data_path)
 
         onnx_model_path = os.path.join(folder_on_server, fn_m[0])
         onnx_model = onnx.load(onnx_model_path)
@@ -256,6 +251,9 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_t
             preds = np.array(predictions[0])
             pred_class = class_name[np.argmax(preds)]
             # get the top most likely results
+            if show_top > len(class_name):
+                show_top = len(class_name)
+            # make sure the top results are ordered most to least likely
             ind = np.array(np.argpartition(preds, -show_top)[-show_top:])
             ind = ind[np.argsort(preds[ind])]
             ind = np.flip(ind)
@@ -263,45 +261,40 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_t
             n_rows = len(top)
             fig = make_subplots(rows=n_rows, cols=3, subplot_titles=("RISE", "KernelShap", "LIME"), row_titles=top,
                                 shared_xaxes=True, vertical_spacing=0.02, horizontal_spacing = 0.02)
-            # To be fixed: define relevances and z specifically for mnist and resnet (+ mirror image for resnet vertically)
-            if class_name == class_name_mnist:
-                relevances_rise = global_store_i('RISE', onnx_model_path, X_test, n_masks=n_masks, feature_res= feature_res, p_keep=p_keep)
+            # check which axis is color channel
+            if X_test.shape[2] <=3:
                 z_rise = X_test[:, :, 0]
+                axis_labels = {2: 'channels'}
+                colorscale='Bluered'
             else:
-                if "RISE" in sel_methods:
-                    relevances_rise = global_store_i('RISE', onnx_model_path, X_test, labels = [range(0, len(class_name))], axis_labels = {0: 'channels'},
-                        n_masks=n_masks, feature_res= feature_res, p_keep=p_keep)
-                    z_rise = np.flipud(X_test[1, :, :])
+                z_rise = X_test[1, :, :]
+                axis_labels = {0: 'channels'}
+                colorscale='jet'
             for m in sel_methods:
                 for i in range(n_rows):
                     if m == "RISE":
                         try:
                             # RISE plot
+                            relevances_rise = global_store_i('RISE', onnx_model_path, X_test, labels = list(range(0, len(class_name))), axis_labels = axis_labels,
+                                n_masks=n_masks, feature_res= feature_res, p_keep=p_keep)
                             fig.add_trace(go.Heatmap(
-                                            z=z_rise, colorscale='gray', showscale=False), i+1, 1)
-                            # To be fixed: color for mnist
-                            if class_name == class_name_mnist:
-                                fig.add_trace(go.Heatmap(
[…]
