### 2022-11-07 d5cefaa
get all methods to work with ImageNet
