### 2022-11-07 f7e4feb
make pylint too-many-branches happy and disble too-many-arguments
