### 2022-11-08 95d2f90
change p_keep autotuning behavior (fixes #380)
