### 2022-11-08 76410ad
Update CITATION.cff
Added Laura.
