### 2022-11-12 #4493 @sara-02
@cwmeijer @diehlpk I have checked the functionality column and was able to able to run the MNIST tutorial given here in the [doc](https://dianna.readthedocs.io/en/latest/tutorials/kernel_mnist.html)
