### 2022-11-12 #4493 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1098/rsta.2019.0054 is OK
> - 10.1145/3328485 is OK
> - 10.1109/ACCESS.2020.3047960 is OK
> - 10.1109/ACCESS.2020.3031477 is OK
> - 10.23919/MIPRO.2018.8400040 is OK
> - 10.2139/ssrn.3503603 is OK
> - 10.5281/zenodo.6303282 is OK
> - 10.5281/zenodo.4724125 is OK
> - 10.5281/zenodo.5012825 is OK
> - 10.5281/zenodo.5061353 is OK
> - 10.18653/v1/P17-1057 is OK
> - 10.7275/extq-7546 is OK
>
> MISSING DOIs
>
> - 10.18653/v1/n16-3020 may be a valid DOI for title: "Why Should I Trust You?": Explaining the Predictions of Any Classifier
> - 10.1007/978-3-642-33709-3_36 may be a valid DOI for title: Leafsnap: A Computer Vision System for Automatic Plant Species Identification
> - 10.18653/v1/p19-1647 may be a valid DOI for title: Symbolic inductive bias for visually grounded learning of spoken language
> - 10.18653/v1/k17-1037 may be a valid DOI for title: Encoding of phonology in a recurrent neural model of grounded speech
> - 10.1613/jair.1.12967 may be a valid DOI for title: Visually grounded models of spoken language: A survey of datasets, architectures and evaluation techniques
>
> INVALID DOIs
>
> - https://doi.org/10.1016/j.jbusres.2020.09.009 is INVALID because of 'https://doi.org/' prefix
> ```

 @elboyran please fix the broken DOI and the missing ones, so we can proceed to publish the paper.
