### 2022-11-15 847f58b
fix 1 doi and add 5 new dois (fix #394)
