+++ paper/paper.bib
@@ -375,7 +375,7 @@
   volume    = {abs/1812.09244},
   year      = {2018},
   url       = {http://arxiv.org/abs/1812.09244},
-  doi       = {10.18653/v1/p19-1647}
+  doi       = {10.18653/v1/p19-1647},
   eprinttype = {arXiv},
   eprint    = {1812.09244},
   timestamp = {Wed, 02 Jan 2019 14:40:18 +0100},
