+++ paper/paper.bib
@@ -410,7 +410,7 @@
   url       = {http://arxiv.org/abs/1706.03815},
   eprinttype = {arXiv},
   eprint    = {1706.03815},
-  doi       = {10.18653/v1/k17-1037}
+  doi       = {10.18653/v1/k17-1037},
   timestamp = {Mon, 13 Aug 2018 16:46:47 +0200},
   biburl    = {https://dblp.org/rec/journals/corr/AlishahiBC17.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
@@ -441,7 +441,7 @@
   url       = {https://arxiv.org/abs/2104.13225},
   eprinttype = {arXiv},
   eprint    = {2104.13225},
-  doi       = {10.1613/jair.1.12967}
+  doi       = {10.1613/jair.1.12967},
   timestamp = {Tue, 04 May 2021 15:12:43 +0200},
   biburl    = {https://dblp.org/rec/journals/corr/abs-2104-13225.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
