### 2022-11-15 a4214a6
remove redundant comment and refactor I add a redundant call to min() for readability. Without this call, the code also runs, but many readers might think it wouldn't. See #380 and PR #389 for a short discussion.
