### 2022-11-15 #4493 @elboyran
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1098/rsta.2019.0054 is OK
> - 10.18653/v1/n16-3020 is OK
> - 10.1145/3328485 is OK
> - 10.1109/ACCESS.2020.3047960 is OK
> - 10.1109/ACCESS.2020.3031477 is OK
> - 10.1016/j.jbusres.2020.09.009 is OK
> - 10.23919/MIPRO.2018.8400040 is OK
> - 10.2139/ssrn.3503603 is OK
> - 10.5281/zenodo.6303282 is OK
> - 10.5281/zenodo.4724125 is OK
> - 10.5281/zenodo.5012825 is OK
> - 10.5281/zenodo.5061353 is OK
> - 10.1007/978-3-642-33709-3_36 is OK
> - 10.18653/v1/p19-1647 is OK
> - 10.18653/v1/P17-1057 is OK
> - 10.18653/v1/k17-1037 is OK
> - 10.7275/extq-7546 is OK
> - 10.1613/jair.1.12967 is OK
>
> MISSING DOIs
>
> - None
>
> INVALID DOIs
>
> - None
> ```

@diehlpk , they are fixed now.
