### 2022-11-15 da96553
organize and finalize coffee model notebook
