+++ example_data/model_generation/coffee/README.md
@@ -1,0 +1,7 @@
+## Coffee model generation
+
+Thе scripts in this folder show how to generate a random forest classifier using sklearn to predict the type of coffee given its spectographs (http://timeseriesclassification.com/description.php?Dataset=Coffee).
+
+The notebooks in this folder are organized as follows:
+
+- [generate_model.ipynb](generate_model.ipynb) defines a model and trains it.
