### 2022-11-15 df9c2cc
Adapt rise to work with 3D data. TODO: make general to work with 2D as well
