### 2022-11-15 9f37f16
Add deeprank run script for RISE
