### 2022-11-17 dianna-ai/dianna#397 @cwmeijer
Create release with matching authors and share on Zenodo for JOSS
Elena wrote (in issue #367):
@cwmeijer and/or @loostrum , can you, please, address thehttps://github.com/openjournals/joss-reviews/issues/4493 by the JOSS editor?
Especially I do not know how can we change our permanent Zenodo record of the releases to match the authors of the paper and if that's possible at all, if we want to do that (SURF people and our PM contributed to one, but not to the other) and do I understand the second point of the editor in the first place? Is it about the filename only containing the authors? What us a DPI? Maybe this release could be then on FigShare? I guess on late September one is not enough because does not contain the latest paper changes.
