### 2022-11-17 dianna-ai/dianna#397 @cwmeijer
I'm guessing DPI is a typo. He means DOI. I can have a look, but especially concerning decisions about authors and such, I would like to do that together.
