+++ dashboard/layouts.py
@@ -253,7 +253,8 @@ def get_uploads_images():
             html.Div([
                 html.Div([
                     dcc.Markdown(children='**Select one/more XAI methods**',
-                        style = {'margin-top': '20px', 'textAlign' : 'center', 'color' : colors['blue1']}
+                        style = {'margin-top': '20px', 'textAlign' : 'center',
+                                'color' : colors['blue1']}
                         ),
                     dcc.Checklist(id = 'method_sel_img',
                         options = [{'label': 'RISE', 'value': 'RISE'},
@@ -271,7 +272,8 @@ def get_uploads_images():
                 ),
                 html.Div([
                     dcc.Markdown(children='**Number of top results to show**',
-                        style = {'margin-top': '20px', 'textAlign' : 'center', 'color' : colors['blue1']}
+                        style = {'margin-top': '20px', 'textAlign' : 'center',
+                                'color' : colors['blue1']}
                         ),
                     dcc.Input(id = 'show_top',
                             type="number",
@@ -291,7 +293,8 @@ def get_uploads_images():
             # Settings bar
             html.Div([
                 html.Div([
-                    html.H6(children='XAI method specific settings', style={'font-weight': 'bold'}),
+                    html.H6(children='XAI method specific settings',
+                            style={'font-weight': 'bold'}),
                 ], className='nine columns'
                 ),
             ],
@@ -301,10 +304,13 @@ def get_uploads_images():
             html.Div([
                 html.Div([
                     dcc.Markdown(children='**Rise**',
-                        style = {'margin-top': '20px', 'textAlign' : 'center', 'color' : colors['blue1']}
+                        style = {'margin-top': '20px', 'textAlign' : 'center',
+                                 'color' : colors['blue1']}
                         ),
                     dcc.Markdown(children='Number of masks',
-                        style = {'margin-left': '5px', 'margin-top': '0px', 'textAlign' : 'left', 'color' : colors['blue1']}
+                        style = {'margin-left': '5px', 'margin-top': '0px',
+                                 'textAlign' : 'left',
+                                 'color' : colors['blue1']}
                         ),
                     dcc.Input(id = 'n_masks',
                         placeholder = "Number of masks",
@@ -318,7 +324,9 @@ def get_uploads_images():
                             }
                     ),
                     dcc.Markdown(children='Feature resolution',
-                        style = {'margin-left': '5px', 'margin-top': '5px', 'textAlign' : 'left', 'color' : colors['blue1']}
+                        style = {'margin-left': '5px', 'margin-top': '5px',
+                                 'textAlign' : 'left',
+                                 'color' : colors['blue1']}
                         ),
                     dcc.Input(id = 'feature_res',
                         placeholder = "Feature res",
@@ -332,7 +340,9 @@ def get_uploads_images():
                             }
                     ),
                     dcc.Markdown(children='Probability to be kept unmasked',
-                        style = {'margin-left': '5px', 'margin-top': '5px', 'textAlign' : 'left', 'color' : colors['blue1']}
+                        style = {'margin-left': '5px', 'margin-top': '5px',
+                                 'textAlign' : 'left',
+                                 'color' : colors['blue1']}
                         ),
                     dcc.Input(id = 'p_keep',
                         placeholder = "P keep",
@@ -348,10 +358,13 @@ def get_uploads_images():
                 ], className = 'three columns'),
                 html.Div([
                     dcc.Markdown(children='**KernelShap**',
-                        style = {'margin-top': '20px', 'textAlign' : 'center', 'color' : colors['blue1']}
+                        style = {'margin-top': '20px', 'textAlign' : 'center',
+                                 'color' : colors['blue1']}
                         ),
[…]
