### 2022-11-17 #4493 @cwmeijer
@diehlpk For various reasons, we have some discrepancy between the list of authors for our current version of the software and the list of authors for the paper, as both the time frames and type work for these 2 works don't fully overlap. Are they required to match? We were not aware of any such requirement. Could you please elaborate?
