+++ setup.cfg
@@ -42,6 +42,7 @@ install_requires =
     onnx-tf
     scikit-image>=0.19.1
     shap
+    spacy
     tensorflow
     tensorflow-probability
     tqdm
