+++ dashboard/callbacks.py
@@ -76,7 +76,8 @@ def upload_image(contents, filename):
             if any(ext in filename[0] for ext in extensions):
                 _, content_string = contents[0].split(',')
 
-                with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
+                with open(os.path.join(folder_on_server, filename[0]), 
+                          'wb') as f:
                     f.write(base64.b64decode(content_string))
 
                 data_path = os.path.join(folder_on_server, filename[0])
@@ -91,15 +92,18 @@ def upload_image(contents, filename):
                     title_x=0.5,
                     title_font_color=layouts.colors['blue1'])
 
-                fig.update_xaxes(showgrid=False, showticklabels=False, zeroline=False)
-                fig.update_yaxes(showgrid=False, showticklabels=False, zeroline=False)
+                fig.update_xaxes(showgrid=False, showticklabels=False,
+                    zeroline=False)
+                fig.update_yaxes(showgrid=False, showticklabels=False,
+                    zeroline=False)
 
                 fig.layout.paper_bgcolor = layouts.colors['blue4']
 
                 return fig
 
             return utilities.blank_fig(
-                text='File format error! <br><br>Please upload only images in one of the following formats:' + extensions)
+                text='File format error! <br><br>Please upload only images in' +
+                    'one of the following formats:' + extensions)
 
         except Exception as e:
             print(e)
@@ -114,14 +118,16 @@ def upload_image(contents, filename):
               dash.dependencies.Input('upload-model-img', 'contents'),
               dash.dependencies.State('upload-model-img', 'filename'))
 def upload_model_img(contents, filename):
-    """Takes in the model file, returns a print statement about its uploading state."""
+    """Takes in the model file, returns a print statement about its uploading
+    state."""
     if contents is not None:
         try:
             if 'onnx' in filename[0]:
 
                 _, content_string = contents[0].split(',')
 
-                with open(os.path.join(folder_on_server, filename[0]), 'wb') as f:
+                with open(os.path.join(folder_on_server, filename[0]),
+                          'wb') as f:
                     f.write(base64.b64decode(content_string))
 
                 return html.Div([f'{filename[0]} uploaded'])
@@ -145,9 +151,11 @@ def upload_model_img(contents, filename):
 # pylint: disable=dangerous-default-value
 # pylint: disable=too-many-arguments
 @cache.memoize()
-def global_store_i(method_sel, model_path, image_test, labels=list(range(2)), axis_labels={2: 'channels'}, 
-        n_masks=1000, feature_res=6, p_keep=.1, n_samples=1000, background=0, n_segments=200, sigma=0, random_state=2):
-    """Takes in the selected XAI method, the model path and the image to test, returns the explainations array."""
+def global_store_i(method_sel, model_path, image_test, labels=list(range(2)),
+    axis_labels={2: 'channels'}, n_masks=1000, feature_res=6, p_keep=.1,
+    n_samples=1000, background=0, n_segments=200, sigma=0, random_state=2):
+    """Takes in the selected XAI method, the model path and the image to test,
+    returns the explainations array."""
     # expensive query
     if method_sel == "RISE":
         relevances = dianna.explain_image(
@@ -180,7 +188,8 @@ def global_store_i(method_sel, model_path, image_test, labels=list(range(2)), ax
      dash.dependencies.State("upload-image", "filename"),
      ])
 def compute_value_i(method_sel, fn_m, fn_i):
-    """Takes in the selected XAI method, the model and the image filenames, returns the selected XAI method."""
+    """Takes in the selected XAI method, the model and the image filenames,
+    returns the selected XAI method."""
[…]
