### 2022-11-17 #4493 @elboyran
@diehlpk what is the reason for this requirement? It does not feel right to include people who  haven't contributed a single line of code  to the software, but had very useful remarks on using the software for a real use case or had useful paper text proof reading and suggestions as well as some of the project management? Also, the paper had been written a while ago when there was a set of developers which has (partly) changed with the new funding and new project.
