### 2022-11-18 #4493 @diehlpk
@elboyran We see the paper and the archive as the same item. It’s essentially a statement that authorship can appear in other ways, that include things other than literal code contributions, and the Zenodo archive (aka, list of contributors) should reflect that as well. If you disagree, I can start a discussion with the editorial board about the topic.
