### 2022-11-21 #4493 @elboyran
> @cwmeijer We had that issue before. In that case, I would advise to make a new release on Zenodo including all authors. In adition, uplaod the same version to FigShare and make a archive for the JOSS paper with having the same authors on the paper. How does that sound?

@diehlpk thank you for your clarifications.

Just to make sure I understand _exactly_ your requirement:

Are you suggesting to make a release on Zenodo where all authors are listed as authors **of the software**
or
a latest release where the paper folder is updated with the very last version and where all authors **of the paper** are listed along with the paper's latest version?
