### 2022-11-21 dianna-ai/dianna#397 @loostrum
@elboyran It would be possible indeed. We could consider creating a JOSS branch and keeping that indefinitely, as an easy way to browse through the code as it was at the time of the paper. If we think that's useful of course.
