### 2022-11-21 dianna-ai/dianna#397 @elboyran
@loostrum yes, lets make a branch and proceed if all authors agree with the proposal.
