### 2022-11-21 dianna-ai/dianna#397 @egpbos
I would say: whatever is the JOSS requirement, fulfill it with the least possible effort.

Btw, if I understand [the cited description here correctly](https://github.com/openjournals/joss-reviews/issues/4493#issuecomment-1320251504), I think Rena would be an edge case, since I don't think she did any "active project direction". It also states that the co-authors all should "agree to be accountable for all aspects of the work". I would say that means that you should really have some knowledge about the code, not just have given some input/feedback in some meetings. In fact, I'm wondering whether even Mathieu from SURF would qualify. He was our prime user and he should probably be mentioned explicitly in the acknowledgements, but he did not co-develop and I would guess he wouldn't like to be held accountable for the code. So, my strict interpretation would be that these authors wouldn't have qualified for this paper. However, I see they are already on the paper draft, so indeed, I would just add them to the software release as well to get the two in sync.

By the way, after the JOSS paper is published, how will we ask people to cite DIANNA? By the most recent software release or the JOSS paper? The most recent release would include all the contributors up to that point as co-authors, which I think is most fair. It's a bit of a shame though that the JOSS paper would not get cited a lot then, although I personally don't mind that much.
