### 2022-11-21 ff44f62
Adapted RISE to generate masks in batches, and possibly to stop early when the results no longer change
