### 2022-11-21 dianna-ai/dianna#397 @elboyran
Part of a comment from @cwmeijer on Teams:

> I think the following might be acceptable (my interpretation of the approach proposed by Elena I guess):

> - Branch a release from the date we submitted the paper, and copy our paper into that branch and release that listing all paper authors as software authors. Downsides: It will require a bit of work and some GIT wizardry (merge cherry picked commits maybe?). We would include a group of people to our author list that haven't contributed anything (significant) and make them accountable for the code as well.
