### 2022-11-21 dianna-ai/dianna#397 @geek-yang
> @cwmeijer , @loostrum is it possible to make a single "JOSS release" with all paper's authors to share on Zenodo and then as of next releases to come back to our current software author's list?
>
> Just checked: given the [history of our releases](https://zenodo.org/search?page=1&size=20&q=conceptrecid:%225592606%22&sort=-version&all_versions=True), it seems more than possible. @geek-yang, @cwmeijer, @loostrum, @egpbos, @gcroci2 do you agree with _adding_ Rena and the SURF people to such a release?

I'm fine with it. and I agree with @egpbos about his concerns about authorship and citation. Indeed I have similar thoughts and it is nice that he summarizes this perfectly 👍!
