### 2022-11-22 ebc9efa
Add note about openmp error to installation instructions
