### 2022-11-22 55d8f70
Use function to automatically determine location of channels axis, function itself is to be implemented
