### 2022-11-22 d37c7bb
Remove axis labels from one test for each image-based method. Tests should once again pass when locate_channels_axis is implemented
