### 2022-11-22 f1c7da2
Implement locate_channels_axis, fix generated axis_labels dict
