### 2022-11-22 72ec25d
Fix another linting issue
