+++ example_data/dataset_preparation/LeafSnap/README.md
@@ -11,7 +11,7 @@ This folder contains 2 notebooks: Data_exploration and Image_cropping.
 - [Data_exploration](Data_exploration.ipynb)
 The purpose of this notebook is to select a subset of the most populous 30 species of lab and field images. Already a [dataset of 30 classes](https://github.com/NLeSC/XAI/blob/master/Software/LeafSnapDemo/Data_preparation_30subset.ipynb) have been selected before, where for the lab images have been cropped semi-manually using IrfanView to remove the riles and color calibration image parts. But 2/3 of that dataset has been selected randomly, not according to the number of images in that class.
 
-This notebook is used to explore the original dataset and find out the most polpulous 30 classes and see which have not been included yet in the previous 30-class dataset.
+This notebook is used to explore the original dataset and find out the most polpulous 30 classes and see which have not been included yet in the previous 20-class dataset.
 
 - [Image_cropping](Image_cropping.ipynb)
 The purpose of this notebook is the croping of the lab images of some species in the LeafSnap30 dataset.
