### 2022-11-24 47b1f38
Update README.md
Updated how to cite.
