+++ setup.cfg
@@ -81,6 +81,7 @@ publishing =
     twine
     wheel
 notebooks =
+    pytorch-lightning
     scipy
     skl2onnx
     spacy
