+++ setup.cfg
@@ -29,6 +29,7 @@ url = https://github.com/dianna-ai/dianna
 version = 0.5.0
 
 [options]
+python_requires = >=3.7,<3.11
 zip_safe = False
 include_package_data = True
 packages = find_namespace:
