### 2022-11-29 7dec865
Fixed the raising of missing image/model/method when not selected issue
