### 2022-11-29 2d84f34
Revert "silent linter"
This reverts commit 5b714d4de6b093cb19ecbb9bc5e4c49defb8f281.
