+++ setup.cfg
@@ -63,6 +63,7 @@ install_requires =
 dev =
     bump2version
     prospector[with_pyroma]
+    pylint==2.15.6
     isort
     pytest
     pytest-cov
