+++ setup.cfg
@@ -62,6 +62,7 @@ install_requires =
 dev =
     bump2version
     prospector[with_pyroma]
+    pylint==2.15.6
     isort
     pytest
     pytest-cov
