### 2022-11-30 dianna-ai/dianna#397 @cwmeijer
I created a branch today from version [v0.4.1](https://github.com/dianna-ai/dianna/releases/tag/v0.4.1) which has as the first line of its description "This is the version at the time of our JOSS paper submission." After some commits and reverting something I messed up my local branch. I'll try again tomorrow. I haven't pushed anything yet.
