### 2022-12-01 2820095
remove unused import, XAI selection methods inline
