+++ dashboard/layouts.py
@@ -2,7 +2,6 @@ import base64
 import utilities
 from dash import dcc
 from dash import html
-import dash_bootstrap_components as dbc
 
 
 #static images
@@ -253,17 +252,20 @@ def get_uploads_images():
             html.Div([
                 html.Div([
                     dcc.Markdown(children='**Select one/more XAI methods**',
-                        style = {'margin-top': '20px', 'textAlign' : 'center',
+                        style = {'margin-top': '20px', "margin-left": "10px",
+                            'textAlign' : 'center',
                             'color' : colors['blue1']}
                         ),
                     dcc.Checklist(id = 'method_sel_img',
                         options = [{'label': 'RISE', 'value': 'RISE'},
                                 {'label': 'KernelSHAP', 'value': 'KernelSHAP'},
                                 {'label': 'LIME', 'value': 'LIME'}],
+                        inline=True,
+                        inputStyle={"margin-left": "10px", "margin-right": "5px"},
                         style={
                                 'margin-left': '0px',
                                 'margin-top': '10px',
-                                'textAlign' : 'left',
+                                'textAlign' : 'center',
                                 'width': '100%',
                                 'color' : colors['blue1']
                             }
