+++ dashboard/callbacks.py
@@ -69,8 +69,8 @@ except Exception:  # If not present, we download
               dash.dependencies.State('upload-image', 'filename'))
 def upload_image(contents, filename):
     """Takes in test image file, returns it as a Plotly figure."""
-    if contents is not None:
 
+    if contents is not None:
         try:
             extensions = ['.png', '.jpg', 'jpeg']
             if any(ext in filename[0] for ext in extensions):
@@ -186,8 +186,8 @@ def global_store_i(method_sel, model_path, image_test, labels=list(range(2)),
      dash.dependencies.State("upload-image", "filename"),
      ])
 def compute_value_i(method_sel, fn_m, fn_i):
-    """Takes in the selected XAI method, the model and the image filenames,
-    returns the selected XAI method."""
+    """Takes in the selected XAI method, the model and the image filenames, returns the selected XAI method."""
+
     if (method_sel is None) or (fn_m is None) or (fn_i is None):
         raise PreventUpdate
 
@@ -231,8 +231,8 @@ def compute_value_i(method_sel, fn_m, fn_i):
 def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image,
     show_top=2, n_masks=1000, feature_res=6, p_keep=0.1, n_samples=1000,
     background=0, n_segments=200, sigma=0, random_state=2):
-    """Takes in the last model and image uploaded filenames, the selected XAI
-    method, and returns the selected XAI method."""
+    """Takes in the last model and image uploaded filenames, the selected XAI method, and returns the selected XAI method."""
+
     ctx = dash.callback_context
 
     if ((ctx.triggered[0]["prop_id"] == "upload-model-img.filename") or 
@@ -245,7 +245,7 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image,
 
     # update graph
     if (fn_m and fn_i) is not None:
-
+        print('fnm and fni not none')
         data_path = os.path.join(folder_on_server, fn_i[0])
         X_test, _ = utilities.open_image(data_path)
 
@@ -382,8 +382,8 @@ def upload_text(clicks, input_value):
               dash.dependencies.Input('upload-model-text', 'contents'),
               dash.dependencies.State('upload-model-text', 'filename'))
 def upload_model_text(contents, filename):
-    """Takes in the model file, returns a print statement about its uploading
-    state."""
+    """Takes in the model file, returns a print statement about its uploading state."""
+
     if contents is not None:
         try:
             if 'onnx' in filename[0]:
@@ -415,8 +415,8 @@ def upload_model_text(contents, filename):
 # and for all time.
 @cache.memoize()
 def global_store_t(method_sel, model_runner, input_text):
-    """Takes in the selected XAI method, the model path and the string to test,
-    returns the explainations highlighted on the string itself."""
+    """Takes in the selected XAI method, the model path and the string to test, returns the explainations highlighted on the string itself."""
+    
     predictions = model_runner(input_text)
     class_name = class_name_text
     pred_class = class_name[np.argmax(predictions)]
@@ -444,8 +444,8 @@ def global_store_t(method_sel, model_runner, input_text):
      dash.dependencies.State("upload-text", "value"),
      ])
 def compute_value_t(method_sel, fn_m, input_text):
-    """Takes in the selected XAI method, the model filename and the text,
-    returns the selected XAI method."""
+    """Takes in the selected XAI method, the model filename and the text, returns the selected XAI method."""
+    
     if (method_sel is None) or (fn_m is None) or (input_text is None):
         raise PreventUpdate
 
@@ -477,8 +477,8 @@ def compute_value_t(method_sel, fn_m, input_text):
 # pylint: disable=too-many-locals
[…]
