### 2022-12-01 7d87702
add paper authors for joss release
