### 2022-12-01 #4493 @elboyran
@diehlpk , we've created the release: https://github.com/dianna-ai/dianna/releases/tag/v0.4.2
