### 2022-12-01 #4493 @diehlpk
Yes, can you please upload the release folder or use Github's Zenodo integration to generate the DOI?
