### 2022-12-01 #4493 @diehlpk
Here is one example on FigShare (https://figshare.com/articles/software/HPX_-_The_C_Standard_Library_for_Parallelism_and_Concurrency/12907034) and here one on Zenodo (https://zenodo.org/record/7377173). Just some random example papers.
