### 2022-12-01 dianna-ai/dianna#397 @elboyran
Any idea why is in not shown in [Zenodo](https://zenodo.org/record/7116103)?
