### 2022-12-01 #4493 @elboyran
Yes, we have Zenodo entry since the first release with all releases since, but this one didn't appear there (yet), will have to find out why.
