### 2022-12-01 #4493 @diehlpk
@elboyran We are close. Thanks for your patience. Can you please change the title on Zenodo to DIANNA: Deep Insight And Neural Network Analysis/ No need for a new release, you can just update the title.
