### 2022-12-01 dianna-ai/dianna#397 @elboyran
Sorry guys @cwmeijer, @loostrum , can you, please, check the editor's comments and maybe edit the release/Zenodo entry? I don't have any time tomorrow and on Monday.

          @elboyran We are close. Thanks for your patience. Can you please change the title on Zenodo to DIANNA: Deep Insight And Neural Network Analysis/ No need for a new release, you can just update the title.

_Originally posted by @diehlpk in https://github.com/openjournals/joss-reviews/issues/4493#issuecomment-1333986182_

        @elboyran Also the order of the authors is different.

_Originally posted by @diehlpk in https://github.com/openjournals/joss-reviews/issues/4493#issuecomment-1334011082_
