### 2022-12-01 #4493 @diehlpk
You go to the archive and click on the orange edit button while logged in to Zenodo.

![image](https://user-images.githubusercontent.com/3922435/205104891-909eb5ea-3221-42ff-8536-2893522d510a.png)
