### 2022-12-01 dianna-ai/dianna#397 @loostrum
Zenodo takes a while to appear, that's why we didn't respond to the editor yet. I do have the edit button, I'll see if I can update the title and author order. We can't change the author order on the Github release btw, as that requires creating a new release. But I think that should be ok.
