### 2022-12-01 dianna-ai/dianna#397 @loostrum
Title and author order fixed, closing this when the editor agrees
