### 2022-12-01 dianna-ai/dianna#397 @elboyran
Thanks!

Get Outlook for Android<https://aka.ms/ghei36>
