### 2022-12-01 #4493 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
