### 2022-12-06 131b2e3
update supported versions in developer instructions
