### 2022-12-06 83d3161
add reference to developer coumentation in readme
