### 2022-12-06 54ab9d2
add run path to test instructions
