### 2022-12-06 b1f2167
make label a required argument (Fixes #131)
