### 2022-12-07 2cc465d
text: add checklist for XAI method selection
