### 2022-12-09 77267cd
Fixed error message with method selection and removed unnecessary run of DIANNA with method selection
