+++ dashboard/callbacks.py
@@ -220,117 +220,117 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_t
         return (html.Div(['Explanation stopped.'], style={'margin-top' : '60px'}),
             utilities.blank_fig())
     # update graph
-    if (ctx.triggered[0]["prop_id"] == "update_button.n_clicks"):
-        if (fn_m and fn_i) is not None and (sel_methods != []):
-            data_path = os.path.join(folder_on_server, fn_i[0])
-            X_test, _ = utilities.open_image(data_path)
+    
+    if (fn_m and fn_i) is not None and (sel_methods != []):
+        data_path = os.path.join(folder_on_server, fn_i[0])
+        X_test, _ = utilities.open_image(data_path)
 
-            onnx_model_path = os.path.join(folder_on_server, fn_m[0])
-            onnx_model = onnx.load(onnx_model_path)
-            # get the output node
-            output_node = prepare(onnx_model, gen_tensor_dict=True).outputs[0]
+        onnx_model_path = os.path.join(folder_on_server, fn_m[0])
+        onnx_model = onnx.load(onnx_model_path)
+        # get the output node
+        output_node = prepare(onnx_model, gen_tensor_dict=True).outputs[0]
 
-            try:
-                predictions = (prepare(onnx_model).run(X_test[None, ...])
-                    [f'{output_node}'])
-                if len(predictions[0]) == 2:
-                    class_name = class_name_mnist
-                else:
-                    class_name = class_names_imagenet
-                # get the predicted class
-                preds = np.array(predictions[0])
-                pred_class = class_name[np.argmax(preds)]
-                # get the top most likely results
-                if show_top > len(class_name):
-                    show_top = len(class_name)
-                # make sure the top results are ordered most to least likely
-                ind = np.array(np.argpartition(preds, -show_top)[-show_top:])
-                ind = ind[np.argsort(preds[ind])]
-                ind = np.flip(ind)
-                top = [class_name[i] for i in ind]
-                n_rows = len(top)
-                fig = make_subplots(rows=n_rows, cols=3,
-                    subplot_titles=("RISE", "KernelShap", "LIME"), row_titles=top,
-                    shared_xaxes=True, vertical_spacing=0.02,
-                    horizontal_spacing = 0.02)
-                # check which axis is color channel
-                if X_test.shape[2] <=3:
-                    z_rise = X_test[:, :, 0]
-                    axis_labels = {2: 'channels'}
-                    colorscale='Bluered'
-                else:
-                    z_rise = X_test[1, :, :]
-                    axis_labels = {0: 'channels'}
-                    colorscale='jet'
-                for m in sel_methods:
-                    for i in range(n_rows):
-                        if m == "RISE":
-                            # RISE plot
-                            relevances_rise = global_store_i('RISE',
-                                onnx_model_path, X_test, labels=[ind[i]],
-                                axis_labels=axis_labels, n_masks=n_masks,
-                                feature_res=feature_res, p_keep=p_keep)
-                            fig.add_trace(
-                                go.Heatmap(z=z_rise, colorscale='gray',
-                                showscale=False), i+1, 1)
-                            fig.add_trace(
-                                    go.Heatmap(z=relevances_rise[0],
-                                        colorscale=colorscale, showscale=False,
-                                        opacity=0.7), i+1, 1)
-                        elif m == "KernelSHAP":
-                            shap_values, segments_slic = global_store_i(
-                                m, onnx_model_path, X_test, labels=[ind[i]],
-                                axis_labels=axis_labels, n_samples=n_samples,
-                                background=background, n_segments=n_segments,
-                                sigma=sigma)
-            
-                            # KernelSHAP plot
-                            fig.add_trace(
-                                go.Heatmap(z=z_rise, colorscale='gray',
[…]
