### 2022-12-14 3e90bb1
add parameter options for text
