### 2022-12-14 cbff077
add some print statements and callback for parameters
