+++ dashboard/callbacks.py
@@ -228,7 +228,8 @@ def compute_value_i(method_sel, fn_m, fn_i):
 # pylint: disable=too-many-locals
 # pylint: disable=unused-argument
 # pylint: disable=too-many-arguments
-def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_top=2, n_masks=1000, feature_res=6, p_keep=0.1, n_samples=1000,
+def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_top=2, 
+    n_masks=1000, feature_res=6, p_keep=0.1, n_samples=1000,
     background=0, n_segments=200, sigma=0, random_state=2, update_button=0, stop_button=0):
     """Takes in the last model and image uploaded filenames, the selected XAI method, and returns the selected XAI method."""
     ctx = dash.callback_context
@@ -367,7 +368,9 @@ def update_multi_options_i(fn_m, fn_i, sel_methods, new_model, new_image, show_t
               dash.dependencies.State('upload-text', 'value'))
 def upload_text(clicks, input_value):
     """Takes in test text string, and print it on the dashboard."""
+    print('in upload text')
     if clicks is not None:
+        print('returning')
         return html.Div([
                     html.P('Input string for the model is:'),
                     html.Br(),
@@ -382,6 +385,7 @@ def upload_text(clicks, input_value):
               dash.dependencies.Input('upload-model-text', 'contents'),
               dash.dependencies.State('upload-model-text', 'filename'))
 def upload_model_text(contents, filename):
+    print('in upload model')
     """Takes in the model file, returns a print statement about its uploading state."""
     if contents is not None:
         try:
@@ -413,8 +417,10 @@ def upload_model_text(contents, filename):
 # redis memory store which is available across processes
 # and for all time.
 @cache.memoize()
-def global_store_t(method_sel, model_runner, input_text):
+def global_store_t(method_sel, model_runner, input_text,
+    n_masks=1000, feature_res=6, p_keep=.1, random_state=2):
     """Takes in the selected XAI method, the model path and the string to test, returns the explainations highlighted on the string itself."""
+    print('in global store t')
     predictions = model_runner(input_text)
     class_name = class_name_text
     pred_class = class_name[np.argmax(predictions)]
@@ -423,14 +429,29 @@ def global_store_t(method_sel, model_runner, input_text):
 
 
     # expensive query
-    relevances = dianna.explain_text(
+    '''    relevances = dianna.explain_text(
         model_runner,
         input_text,
         tokenizer,
         method_sel,
         labels=[pred_idx]
-        )
+        )'''
+    if method_sel == "RISE":
+        print('calculare rel rise')
+        relevances = dianna.explain_text(
+            model_runner, input_text, tokenizer,
+            method=method_sel,
+            labels=[pred_idx],
+            n_masks=n_masks, feature_res=feature_res, p_keep=p_keep)
 
+    else:
+        relevances = dianna.explain_text(
+            model_runner, input_text, tokenizer,
+            'LIME',
+            random_state=random_state,
+            labels=[pred_idx],
+            preprocess_function=utilities.preprocess_function)
+    print('calculated rel')
     return relevances
 
 
@@ -442,6 +463,7 @@ def global_store_t(method_sel, model_runner, input_text):
      dash.dependencies.State("upload-text", "value"),
      ])
 def compute_value_t(method_sel, fn_m, input_text):
+    print('in compute value t')
     """Takes in the selected XAI method, the model filename and the text, returns the selected XAI method."""
     if (method_sel is None) or (fn_m is None) or (input_text is None):
         raise PreventUpdate
[…]
